% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actionbuttonUI.R
\name{actionButtonUI}
\alias{actionButtonUI}
\title{Wrapper around shiny::numericInput.}
\usage{
actionButtonUI(id,
  label = paste("Please provide a label argument to the actionButton",
  " with id = ", id, ".", sep = ""), icon = NULL, width = NULL,
  custom_css = NULL)
}
\arguments{
\item{id}{A unique string id for the numericInput. If your code is modularised,
remember to wrap the id with the relevant session namespace function, e.g.
numericValueUI(id = session$ns("someinput")).}

\item{icon}{An icon for the action button (see shiny::icon).}

\item{width}{The width of the shiny::numericValue UI element. Must be valid
css (i.e. "10%" or "150px" will work).}

\item{custom_css}{If provided, the action button is wrapped in a div styled 
by the custom_css input. For example, one could provide 
custom_css = "display: inline-block;" to display the action button inline.}

\item{width}{The widths of the action buttons. Must be valid css.}
}
\description{
Wrapper around shiny::numericInput.
}
