% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actionbuttonUI.R
\name{actionButtonUIs}
\alias{actionButtonUIs}
\title{Creates a tagList of action buttons wrapped in a (potentially styled) div.}
\usage{
actionButtonUIs(ids, labels = NULL, icons = NULL, widths = NULL,
  custom_css = NULL)
}
\arguments{
\item{ids}{A unique string id for the numericInput. If your code is modularised,
remember to wrap the id with the relevant session namespace function, e.g.
numericValueUI(id = session$ns("someinput")).}

\item{labels}{A label for the input.}

\item{icons}{Icons for the action buttons.}

\item{widths}{The widths of the action buttons. Must be valid
css (i.e. "10%" or "150px" will work).}

\item{custom_css}{If provided, the action buttons are wrapped in divs each
styled by the custom_css input. For example, when creating multiple action
buttons, one could provide custom_css = "display: inline-block;" to display 
the action buttons inline. To use the same style on all action buttons, 
provide a single-text string or single-element vector. To use different
styles for each action button, provide a vector of custom css text strings.}
}
\description{
Creates a tagList of action buttons wrapped in a (potentially styled) div.
}
\examples{
actionButtonUIs(ids = c("trigger_calculation", "trigger_plot"))
actionButtonUIs(ids = c("trigger_calculation", "trigger_plot"), 
custom_css = "display: inline-block")
actionButtonUIs(ids = c("trigger_calculation", "trigger_plot"), 
custom_css = c("display: inline-block", "display: inline-flex")
}
