% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action.R
\name{actionServer}
\alias{actionServer}
\title{Action Button Server}
\usage{
actionServer(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}
}
\value{
The return value, if any, from executing the module server function
}
\description{
Create an action button server.
}
\details{
\preformatted{
## Example
actionServer("action")
}
}
\seealso{
\code{\link[=actionUI]{actionUI()}}
}
