% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_data.R
\name{add_temp_var}
\alias{add_temp_var}
\title{Add temporal (hour, day, ...) using a date column}
\usage{
add_temp_var(
  data,
  col_date,
  tz = "UTC",
  temp_vars = c("min", "heure", "jour_semaine", "semaine_annee", "mois_annee", "annee")
)
}
\arguments{
\item{data}{\code{data.table}. Table containing a date colummn.}

\item{col_date}{\code{character}. Name of the date column.}

\item{tz}{\code{character}. Timezone of new columns. Default to "UTC"}

\item{temp_vars}{\code{character} (c("min", "heure", "jour_semaine", "semaine_annee", "mois_annee", "annee")). Variables to be added to the table.}
}
\value{
a data.table with new columns.
}
\description{
Add temporal (hour, day, ...) using a date column
}
\examples{
\dontrun{\donttest{

data <- data.table(date = seq(Sys.time(), Sys.time() + 24*60*60, length.out = 100))

data <- add_temp_var(data = data,
                     col_date = "date",
                     tz = "CET",
                     temp_vars = c("heure", "semaine_annee", "annee"))
                 
data_utc <- add_temp_var(data = data,
                     col_date = "date",
                     tz = "UTC",
                     temp_vars = c("heure", "semaine_annee", "annee"))
                     
}}
}
