% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkboxGroupButtons.R
\name{checkboxGroupButtonsServer}
\alias{checkboxGroupButtonsServer}
\title{Checkbox group buttons Server}
\usage{
checkboxGroupButtonsServer(id, status = "primary", choices, selected = NULL)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{status}{Add a class to the buttons, you can use Bootstrap status like
'info', 'primary', 'danger', 'warning' or 'success'. Or use an arbitrary
strings to add a custom class, e.g. : with status = "custom-class",
buttons will have class btn-custom-class.}

\item{choices}{List of values to select from.}

\item{selected}{The initially selected value.}
}
\value{
The return value, if any, from executing the module server function
}
\description{
Create a checkbox group buttons server.
}
\details{
\preformatted{
## Example
checkboxGroupButtonsServer(
  id = "checkboxgroupbuttons", label = "",
  status = "primary", choices = c("A", "B", "C"), selected = "A")
}
}
\seealso{
\code{\link[=checkboxGroupButtonsUI]{checkboxGroupButtonsUI()}}
}
