% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkboxGroupButtons.R
\name{checkboxGroupButtonsUI}
\alias{checkboxGroupButtonsUI}
\title{Checkbox Group buttons UI}
\usage{
checkboxGroupButtonsUI(
  id,
  label = "",
  status = "primary",
  choices = as.character(0),
  selected = as.character(0)
)
}
\arguments{
\item{id}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}

\item{status}{Add a class to the buttons, you can use Bootstrap status like
'info', 'primary', 'danger', 'warning' or 'success'. Or use an arbitrary
strings to add a custom class, e.g. : with status = "custom-class",
buttons will have class btn-custom-class.}

\item{choices}{List of values to select from.}

\item{selected}{The initially selected value.}
}
\value{
The return value, if any, from executing the module server function
}
\description{
Create checkbox group buttons UI.
}
\details{
\preformatted{
## Example
checkboxGroupButtonsUI(
  id = "checkboxgroupbuttons", label = "",
  status = "primary")
}
}
\seealso{
\code{\link[=checkboxGroupButtonsServer]{checkboxGroupButtonsServer()}}
}
