% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkboxGroup.R
\name{checkboxGroupServer}
\alias{checkboxGroupServer}
\title{Checkbox group Server}
\usage{
checkboxGroupServer(id, choices, selected = NULL, reverse = FALSE)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{choices}{List of values to select from.}

\item{selected}{The initially selected value.}

\item{reverse}{A boolean value whether to reverse the choices or not.}
}
\value{
The return value, if any, from executing the module server function
}
\description{
Create a checkbox group server.
}
\details{
\preformatted{
## Example
checkboxGroupServer("checkboxgroup", choices, selected, reverse = FALSE)
}
}
\seealso{
\code{\link[=checkboxGroupUI]{checkboxGroupUI()}}
}
