% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkboxGroup.R
\name{checkboxGroupUI}
\alias{checkboxGroupUI}
\title{Checkbox Group UI}
\usage{
checkboxGroupUI(
  id,
  label = NULL,
  choices = character(0),
  selected = character(0),
  value = FALSE,
  inline = FALSE
)
}
\arguments{
\item{id}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}

\item{choices}{List of values to select from.}

\item{selected}{The initially selected value.}

\item{value}{A boolean value for All/None value (All: TRUE, None: FALSE)}

\item{inline}{If TRUE, render the choices inline (i.e. horizontally)}
}
\value{
A checkbox group UI
}
\description{
Create a checkbox group UI.
}
\details{
\preformatted{
## Example
checkboxGroupUI("checkboxgroup")
}
}
\seealso{
\code{\link[=checkboxGroupServer]{checkboxGroupServer()}}
}
