% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkboxUI.R
\name{checkboxUI}
\alias{checkboxUI}
\title{checkboxUI}
\usage{
checkboxUI(id, label = paste("Please provide a label argument to the",
  "checkbox with id = ", id, ".", sep = ""), icon = NULL, width = NULL,
  custom_css = NULL)
}
\arguments{
\item{id}{A unique string id for the checkbox. If your code is modularised,
remember to wrap the id with the relevant session namespace function, e.g.
checkboxUI(id = session$ns("someinput")).}

\item{label}{A label for the checkbox.}

\item{width}{The width of the entire shiny::checkbox UI element. Must be 
valid css (i.e. '10\%' or '150px' will work).}

\item{custom_css}{If provided, the checkbox is wrapped in a div styled 
by the custom_css input. For example, one could provide 
custom_css = "display: inline-block;" to display the checkbox inline.}
}
\description{
Wrapper around shiny::checkboxInput
}
