% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkboxUI.R
\name{checkboxUIs}
\alias{checkboxUIs}
\title{checkboxUIs}
\usage{
checkboxUIs(ids, labels = NULL, icons = NULL, widths = NULL,
  custom_css = NULL)
}
\arguments{
\item{ids}{String id for the checkboxex. If your code is modularised,
remember to wrap the id with the relevant session namespace function, e.g.
checkboxUIs(id = sapply(c("someinput", "anotherinput", session$ns))).}

\item{labels}{Labels for the checkboxes.}

\item{widths}{The widths of the checkboxes. Must be valid css (i.e. 
"10\%" or "150px" will work; see shiny::validateCssUnit). To use the same 
width on all checkboxes, provide a single-text string or 
a vector with one text string. To use different widths for each checkbox,
provide a vector of custom css text strings with one element per 
action button.}

\item{custom_css}{If provided, the  checkboxes are wrapped in divs each
styled by the custom_css input. For example, when creating multiple 
checkboxes one could provide custom_css = "display: inline-block;" to 
display them inline. To use the same style on all checkboxes, 
provide a single-text string or single-element vector. To use different
styles for each checkbox, provide a vector of custom css text strings
with one element per checkbox.}
}
\description{
Wrapper around shiny::checkboxInput
}
