% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{clusterApp}
\alias{clusterApp}
\title{A test shiny app for the cluster module}
\usage{
clusterApp()
}
\value{
a [shiny::shinyApp()] object.
}
\description{
`clusterApp()` creates a small test app for testing the [clusterInput()] and
[clusterServer()] functions. It uses a subset of the package dataset `counts`,
clusters the counts based on the selected check boxes and displays them
as a table.
}
\examples{

clusterApp()

}
