% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{clusterServer}
\alias{clusterServer}
\title{Server function to cluster a count matrix by rows and/or columns}
\usage{
clusterServer(id, counts = NULL, gene_metadata = NULL, debug = FALSE)
}
\arguments{
\item{id}{namespace id for the UI components. Must match the id provided to the
[clusterInput()] function.}

\item{counts}{a reactive counts object. Should contain only numeric columns}

\item{gene_metadata}{a reactive object. Contains the metadata for the genes
present in the counts object.}

\item{debug}{Turn on debugging message statements}
}
\value{
a [shiny::reactive()] object which is the clustered counts
}
\description{
`clusterServer()` clusters the supplied counts either by rows or columns
or both
}
\examples{

clusterServer("rnaseqData", counts = reactive(rnaseqtools::counts[1:10,1:5]))

}
