% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_heatmap.R
\name{colorData}
\alias{colorData}
\title{Associeted colors to data.frame}
\usage{
colorData(data, nbclasses = NULL, col = c("#FF0000", "#FFFFFF", "#0000FF"),
  colorby = "all")
}
\arguments{
\item{data}{: data.frame}

\item{nbclasses}{: number of classes}

\item{col}{: 3 col to use in colorRampPalette}

\item{colorby}{: can be "all","row","col".}
}
\value{
data.frame compound to original data.frame and associated color data.frame
}
\description{
Associeted colors to data.frame
}
