% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commentsModule.R
\name{commentsServer}
\alias{commentsServer}
\title{the server for a comments shiny module}
\usage{
commentsServer(input, output, session, dbConnection, entityName)
}
\arguments{
\item{input}{enviroment provide by shiny}

\item{output}{enviroment provide by shiny}

\item{session}{enviroment provide by shiny}

\item{dbConnection}{an sqlite database connetion}
}
\description{
the server for a comments shiny module
}
