% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commentsModule.R
\name{commentsUI}
\alias{commentsUI}
\title{the UI for a persistent comments shiny module}
\usage{
commentsUI(id)
}
\arguments{
\item{id}{the html document's widget id}
}
\description{
the UI for a persistent comments shiny module
}
