% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commentsWithBrowseModule.R
\name{commentsWithBrowseServer}
\alias{commentsWithBrowseServer}
\title{the server for a commentsWithBrowse shiny module}
\usage{
commentsWithBrowseServer(input, output, session, dbConnection, entityName)
}
\arguments{
\item{input}{enviroment provide by shiny}

\item{output}{enviroment provide by shiny}

\item{session}{enviroment provide by shiny}

\item{dbConnection}{an sqlite database connetion}
}
\description{
the server for a commentsWithBrowse shiny module
}
