% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commentsWithBrowseModule.R
\name{commentsWithBrowseUI}
\alias{commentsWithBrowseUI}
\title{the UI for a persistent commentsWithBrowse shiny module}
\usage{
commentsWithBrowseUI(id)
}
\arguments{
\item{id}{the html document's widget id}
}
\description{
the UI for a persistent commentsWithBrowse shiny module
}
