% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_indicators.R
\name{compute_idc}
\alias{compute_idc}
\title{Compute indicators for given model and data}
\usage{
compute_idc(object = NULL, data, col_obs, col_fit = NULL, by = NULL,
  dec = 2, nb.cores = 1, label_no_choice = "Aucun")
}
\arguments{
\item{object}{\code{modelInstant} (NULL). A modelInstant object. IE = ???}

\item{data}{\code{data.frame} / \code{data.table}. Data on which to measure the quality of the model.}

\item{col_obs}{\code{character}. Column name of observed values.}

\item{col_fit}{\code{character} (NULL). Column name of fitted values.}

\item{by}{\code{character} (NULL). Column name of aggregation variable.}

\item{dec}{\code{integer} (2). Number of decimals to be kept.}

\item{nb.cores}{\code{integer} (1). Number of cores, when oject is used to make prediction.}

\item{label_no_choice}{\code{character} ("Aucun"). Label for no selection.}
}
\value{
a data.frame whose columns are the required indicators and aggregation column.
}
\description{
Compute indicators for given model and data
}
\examples{
\dontrun{\donttest{

library(data.table)

obs <- round(runif(100, 1, 10))
fit <- rnorm(100, 1, 5)

compute_idc(data = data.table(obs = obs, fit = fit), 
col_obs = "obs", col_fit = "fit", by = "Aucun")
compute_idc(data = data.table(obs = obs, fit = fit, by_var = sample(rep(1:10, 10))), 
col_obs = "obs", col_fit = "fit", by = "by_var")

}}
}
