% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_data.R
\name{compute_model_idc}
\alias{compute_model_idc}
\title{Compute indicators for given model and data}
\usage{
compute_model_idc(
  object = NULL,
  data,
  col_obs,
  col_fit = NULL,
  by = NULL,
  dec = 2,
  label_no_choice = "None",
  ...
)
}
\arguments{
\item{object}{\code{model object} (NULL). Can be a R model. So we compute prediction on \code{data} using 
\code{fit = predict(object, newdata = data, ...)}. Else, \code{col_fit} must be in \code{data}}

\item{data}{\code{data.frame} / \code{data.table}. Data on which to measure the quality of the model.}

\item{col_obs}{\code{character}. Column name of observed values.}

\item{col_fit}{\code{character} (NULL). Column name of fitted values.}

\item{by}{\code{character} (NULL). Column name of aggregation variable.}

\item{dec}{\code{integer} (2). Number of decimals to be kept.}

\item{label_no_choice}{\code{character} ("None"). Label for no selection.}

\item{...}{in case of \code{model object}, optionnals arguments of \code{predict}}
}
\value{
a data.frame whose columns are the required indicators and aggregation column.
}
\description{
Compute indicators for given model and data
}
\examples{
\dontrun{\donttest{

library(data.table)

obs <- round(runif(100, 1, 10))
fit <- rnorm(100, 1, 5)

compute_model_idc(data = data.table(obs = obs, fit = fit), 
    col_obs = "obs", col_fit = "fit", by = "None"
)

data_idc <- compute_model_idc(
   data = data.table(obs = obs, fit = fit, by_var = sample(rep(1:10, 10))), 
   col_obs = "obs", col_fit = "fit", by = "by_var"
)
plot_idc_table(data_idc = data_idc)

}}
}
