% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_heatmap.R
\name{constructdata}
\alias{constructdata}
\title{Associeted constructor data.frame to initial data.frame}
\usage{
constructdata(data, rownames)
}
\arguments{
\item{data}{: data.frame}

\item{rownames}{: Rownames of the heatmap graphic}
}
\value{
data.frame compound to original data.frame 
and associated constructor data.frame
}
\description{
Associeted constructor data.frame to initial data.frame
}
