% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pie.R, R/countPlot.R
\name{countPlotApp}
\alias{countPlotApp}
\title{A test shiny app for the countPlot module}
\usage{
countPlotApp(debug = TRUE)

countPlotApp(debug = TRUE)
}
\value{
a [shiny::shinyApp()] object.

a [shiny::shinyApp()] object.
}
\description{
`countPlotApp()` creates a small test app for testing the [countPlotOutput()] and
[countPlotServer()] functions. It uses a subset of the package datasets `counts`,
`sampleInfo` and `gene_metadata` and create a [ggplot2::ggplot()] countPlot object.
It also has transform radio buttons to test using that information to change
the colour palette.

`countPlotApp()` creates a small test app for testing the [countPlotOutput()] and
[countPlotServer()] functions. It uses a subset of the package datasets `counts`,
`sampleInfo` and `gene_metadata` and create a [ggplot2::ggplot()] countPlot object.
It also has transform radio buttons to test using that information to change
the colour palette.
}
\examples{

countPlotApp()


countPlotApp()

}
