% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pie.R, R/countPlot.R
\name{countPlotInput}
\alias{countPlotInput}
\title{Create UI components to display a pie chart}
\usage{
countPlotInput(id)

countPlotInput(id)
}
\arguments{
\item{id}{namespace id for the UI components. Must match the id provided to the
[countPlotInput()] and [countPlotOutput()] functions.}
}
\value{
a [htmltools::tagList()] containing a [shiny::selectInput()] object

a [htmltools::tagList()] containing a [shiny::selectInput()] object
}
\description{
`PieInput()` produces a [shiny::selectInput()] control for choosing a gene
to display as a count plot.

`countPlotInput()` produces a [shiny::selectInput()] control for choosing a gene
to display as a count plot.
}
\examples{

countPlotOutput("rnaseqData")


countPlotOutput("rnaseqData")

}
