% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pie.R, R/countPlot.R
\name{countPlotOutput}
\alias{countPlotOutput}
\title{Create UI components to display a count plot for a selected gene}
\usage{
countPlotOutput(id)

countPlotOutput(id)
}
\arguments{
\item{id}{namespace id for the UI components. Must match the id provided to the
[countPlotServer()] function.}
}
\value{
a [htmltools::tagList()] containing a [shiny::plotOutput] object

a [htmltools::tagList()] containing a [shiny::plotOutput] object
}
\description{
`countPlotOutput()` produces a plotOutput space for a plot

`countPlotOutput()` produces a plotOutput space for a plot
}
\examples{

countPlotOutput("rnaseqData")


countPlotOutput("rnaseqData")

}
