% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countXYScatterPlot.R
\name{countXYScatterPlotApp}
\alias{countXYScatterPlotApp}
\title{A test shiny app for the countXYScatterPlot module}
\usage{
countXYScatterPlotApp(debug = TRUE, hide_x = FALSE)
}
\value{
a [shiny::shinyApp()] object
}
\description{
`countXYScatterPlotApp()` creates a small test app for testing the [countXYScatterPlotInput()] and
[countXYScatterPlotServer()] functions. A subset of the returned sample and count data.frames
are displayed in two [shiny::tableOutput()]s
}
\examples{
countXYScatterPlotApp()
}
