% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countXYScatterPlot.R
\name{countXYScatterPlotInput}
\alias{countXYScatterPlotInput}
\title{Create UI components to select two genes to plot against each other}
\usage{
countXYScatterPlotInput(id, x_label = "Gene1 X:", y_label = "Gene2 Y:")
}
\arguments{
\item{id}{namespace id for the UI components. Must match the id provided to the
[countXYScatterPlotServer()] function.}

\item{x_label}{character Label for X variable select input}

\item{y_label}{character Label for Y variable select input}
}
\value{
a [htmltools::tagList()] containing two [shiny::selectizeInput()]
controls and one [shiny::checkboxInput()]
}
\description{
`countXYScatterPlotInput()` produces the select inputs for selecting the
variables to plot against each other
}
\examples{

countXYScatterPlotInput("countData")

}
