% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countXYScatterPlot.R
\name{countXYScatterPlotOutput}
\alias{countXYScatterPlotOutput}
\title{Create Output to display the Scatterplot}
\usage{
countXYScatterPlotOutput(id)
}
\arguments{
\item{id}{namespace id for the UI components. Must match the id provided to the
[countXYScatterPlotServer()] function.}
}
\value{
a [htmltools::tagList()] containing three [shiny::plotOutput()]s
}
\description{
`countXYScatterPlotOutput()` produces a PlotOutput element for the scatterplot
}
\examples{

countXYScatterPlotOutput("countData")

}
