% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countXYScatterPlot.R
\name{countXYScatterPlotServer}
\alias{countXYScatterPlotServer}
\title{Server function to produce an xy scatter plot}
\usage{
countXYScatterPlotServer(
  id,
  counts = NULL,
  sample_info = NULL,
  gene_metadata = NULL,
  debug = FALSE
)
}
\arguments{
\item{id}{namespace id for the UI components. Must match the id provided to the
[countXYScatterPlotInput()] function.}

\item{counts}{a reactive counts object. Should contain only numeric columns}

\item{sample_info}{a reactive object. Represents the samples and associated
metadata}

\item{gene_metadata}{a reactive object. Contains the metadata for the genes
present in the counts object.}
}
\value{
a list containing two [shiny::reactive()] objects
* sampleInfo a data.frame of sample metadata
* counts a data.frame of RNAseq count data
}
\description{
`countXYScatterPlotServer()` takes a data frame of count data and produces
a scatterplot based on the genes selected
}
\examples{

countXYScatterPlotServer("rnaseqData")

}
