% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTableModule.R
\name{dataTableServer}
\alias{dataTableServer}
\title{the server for a DataTable shiny module}
\usage{
dataTableServer(id, input, output, session, tbl, selectionPolicy)
}
\arguments{
\item{input}{enviroment provided by shiny}

\item{output}{enviroment provided by shiny}

\item{session}{enviroment provided by shiny}

\item{tbl}{data.frame}

\item{selectionPolicy}{character string, single, multiple or none}

\item{selectionMechanism}{character string, manual or programmatic}

\item{pageLength}{integer typically 5, 10, 25 or 50}

\item{visibleRows}{"all", "none", or a list of rownames}

\item{selectedRows}{character string, reactive(NULL)  by default}

\item{searchTerm}{character string, "" by default}

\item{wrapLongTextInCells}{logcial default TRUE}
}
\description{
the server for a DataTable shiny module
}
