% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTableModule.R
\name{dataTableUI}
\alias{dataTableUI}
\title{the UI for a DataTable shiny module}
\usage{
dataTableUI(id)
}
\arguments{
\item{id}{the html document's widget id}
}
\description{
the UI for a DataTable shiny module
}
