% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{dateServer}
\alias{dateServer}
\title{Date Server}
\usage{
dateServer(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}
}
\value{
The return value, if any, from executing the module server function.
}
\description{
Create a date server.

\preformatted{
## Example
  dateServer("date")
}
}
\seealso{
\code{\link[=dateUI]{dateUI()}}
}
