% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{dateUI}
\alias{dateUI}
\title{Date UI}
\usage{
dateUI(id, label = "Date", value = Sys.Date(), format = "yyyy-mm-dd")
}
\arguments{
\item{id}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}

\item{value}{A date value.}

\item{format}{A date format value.}
}
\value{
A date UI
}
\description{
Create a date UI.

\preformatted{
## Example
dateUI("date", label = "Date", value = Sys.Date(), format = "yyyy-mm-dd")
}
}
