% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_heatmap.R
\name{.amHeatmap}
\alias{.amHeatmap}
\title{Amchart Heat-Map}
\usage{
.amHeatmap(
  data,
  nbclasses = 5,
  col = c("#FF0000", "#FFFFFF", "#0000FF"),
  labels = TRUE,
  cex = 10,
  main = "",
  rownames,
  xLabelsRotation = 45,
  colorby = "all",
  legend = TRUE
)
}
\arguments{
\item{data}{: data.frame, should be a contingency table}

\item{nbclasses}{: number of classes}

\item{col}{: 3 col to use in colorRampPalette}

\item{labels}{: TRUE FALSE, display labels}

\item{cex}{: size of labels}

\item{main}{: title}

\item{rownames}{: Rownames of the heatmap graphic}

\item{xLabelsRotation}{: rotation of xlabels}

\item{colorby}{: can be "all","row","col".}

\item{legend}{: TRUE or FALSE, display legend}
}
\value{
data.frame compound to original data.frame and associated constructor data.frame
}
\description{
Amchart Heat-Map
}
\examples{
\dontrun{
data(USArrests, "VADeaths")
USArrests <- USArrests [1:10,]
amHeatmap(USArrests, rownames = rownames(USArrests))
amHeatmap(USArrests, nbclasses=5, col = c("#FF0000", "#FFFFFF", "#0000FF"),
labels = TRUE, cex = 10, main="My title", xLabelsRotation = 45, colorby = "all",
legend = TRUE, rownames = rownames(USArrests))
amHeatmap(USArrests, nbclasses = 5, col = c("#FF0000","#FFFFFF","#0000FF"),
labels = TRUE, cex = 10, main = "My title",xLabelsRotation = 45,
colorby="row",legend = TRUE, rownames = rownames(USArrests))
amHeatmap(USArrests, nbclasses = 5, col = c("#FF0000", "#FFFFFF", "#0000FF"),
labels = TRUE, cex = 10, main = "My title", xLabelsRotation = 45,
colorby="col",legend = TRUE, rownames = rownames(USArrests))
amHeatmap(USArrests, nbclasses=10, col=c("#00FF00","#FF00FF","#0000FF"),
labels = TRUE, cex = 10, main = "My title", xLabelsRotation = 45, 
colorby="all",legend = TRUE, rownames = rownames(USArrests))
}
}
