% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicators.R
\name{.get_dates_indicators}
\alias{.get_dates_indicators}
\title{get stats indicator on dates data}
\usage{
.get_dates_indicators(data, dates_vars, optional_stats = "all",
  keep_dataframe = TRUE, keep_datatable = TRUE)
}
\arguments{
\item{data}{\code{data.frame}, \code{data.table}}

\item{dates_vars}{\code{character}}

\item{optional_stats}{\code{character} optional statistics computed on data,
you can look at \link{show_data} for more information.}

\item{keep_dataframe}{\code{logical} data.frame in output ?}

\item{keep_datatable}{\code{logical} datatable (DT) in output ?}
}
\value{
DT with statistics on dates data
}
\description{
return statistics from dates data, is called by 
\link{get_dt_num_dt_fac}
}
