% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicators.R
\name{.get_stats_info}
\alias{.get_stats_info}
\title{Add information on variables in DT table}
\usage{
.get_stats_info(cols, vars_infos, nb_modal2show = NULL)
}
\arguments{
\item{cols}{\code{character}. Columns names the table to be displayed.}

\item{vars_infos}{\code{character}. Table with cols 'name' and 'info'.}

\item{nb_modal2show}{\code{integer}. todo.}
}
\value{
DT container arg.
}
\description{
Add information on variables in DT table
}
\examples{
\dontrun{\donttest{

library(shiny)

# create table of variables informations
vars_infos <- data.table(
  # names of all possible variables 
  name = c("Sepal.Length", "Species", "Petal.Length", "Sepal.Width", "Petal.Width"), 
  # corresponding infos
  info = c("Infos about Sepal.Length", "Infos about Species", "Infos about Petal.Length", 
           "Infos about Sepal.Width", "Infos about Petal.Width"))

# shiny app
ui <- fluidPage(
  fluidRow(
    column(12,
           DT::dataTableOutput("my_DT_table")        
    )
  )
)

server <- function(input, output, session) {
  output$my_DT_table <- DT::renderDataTable({
    DT::datatable(
      data.table(iris), 
      
      rownames = F, filter = 'bottom', 
      
      container = .get_stats_info(
        cols = names(iris),
        vars_infos = vars_infos
      )
    )
  })
}

shinyApp(ui, server)

}}
}
