% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plots.R
\name{.plotTimeSeries}
\alias{.plotTimeSeries}
\title{Plot time series}
\usage{
.plotTimeSeries(
  data,
  col.date,
  col.series,
  ylab = "values",
  main = NULL,
  aggregation = "Average",
  col = c("red", "black"),
  point.size = 5,
  maxSeries = 300,
  js = TRUE
)
}
\arguments{
\item{data}{: data}

\item{col.date}{: column name for data values}

\item{col.series}{: column(s) name for serie(s)}

\item{ylab}{: label of the numeric variable}

\item{main}{: Title}

\item{aggregation}{: Javascript only. one of 'Average' (Default), 'Sum', 'Low', 'High'}

\item{col}{: color}

\item{point.size}{: point size, default to 5}

\item{maxSeries}{: Javascript only. maximum number of points 
before aggregation. Default to 300}

\item{js}{: boolean. TRUE (Default) = javascript plot, FALSE = base plot}
}
\description{
Plot time series

This function can plot timeseries in a dynamic or static way
}
\examples{
\dontrun{
data <- data.frame(date = seq(ISOdate(2000,1,1), by = "10 min", length.out = 500), 
 conso = rnorm(500), fitted.values = rnorm(500))
 
plotTimeSeries(data, col.date = "date", col.series = "conso", main = "Consommation")

plotTimeSeries(data, col.date = "date", col.series = "conso", main = "Consommation", js = FALSE)

plotTimeSeries(data, col.date = "date", col.series = "conso",
 aggregation = "Sum")
 
plotTimeSeries(data, col.date = "date", col.series = c("conso", "fitted.values"),
 col = c("blue", "purple"), point.size = 2)
 
plotTimeSeries(data, col.date = "date", col.series = c("conso", "fitted.values"),
 col = c("blue", "purple"), point.size = 2, js = FALSE)
 }
 

}
