% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_indicators.R
\name{.rmse}
\alias{.rmse}
\title{Indicators funs}
\usage{
.rmse(real, fit, na.rm = TRUE)
}
\arguments{
\item{real}{\code{numeric}. Observed values.}

\item{fit}{\code{numeric}. Fitted values.}

\item{na.rm}{\code{boolean}. Whether to ignore NA values.}
}
\value{
a  numeric. The value of the called indicator.
}
\description{
Indicators funs
}
\examples{
\dontrun{\donttest{

real <- round(runif(100, 1, 10))
fit <- rnorm(100, 1, 5)

mape <- .mape(real, fit)
mape_e <- .mape_e(real, fit)
mae <- .mae(real, fit)
rmse <- .rmse(real, fit)

}}
}
