% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{downloadDataServer}
\alias{downloadDataServer}
\title{Data Download Server}
\usage{
downloadDataServer(id, filename, data)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{filename}{A string specifying a file name}

\item{data}{A data object}
}
\value{
The return value, if any, from executing the module server function
}
\description{
Create data download server

\preformatted{
  downloadDataServer("download_data", filename = "tmp", data = data)
}
}
\seealso{
\code{\link[=downloadDataUI]{downloadDataUI()}}
}
