% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{downloadDataUI}
\alias{downloadDataUI}
\title{Data Download Button UI}
\usage{
downloadDataUI(id, label, right = FALSE, up = TRUE, animate = TRUE)
}
\arguments{
\item{id}{The name of the output slot that the shiny::downloadHandler is assigned to.}

\item{label}{The label that should appear on the button.}

\item{right}{A boolean specifying right or not.}

\item{up}{A boolean specifying up or not.}

\item{animate}{A boolean whether to use animation or not.}
}
\value{
A data download button UI
}
\description{
Create a data download button UI.

\preformatted{
  downloadDataUI("download_data", label = "download")
}
}
\seealso{
\code{\link[=downloadDataServer]{downloadDataServer()}}
}
