% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{downloadPlotServer}
\alias{downloadPlotServer}
\title{Plot Download Server}
\usage{
downloadPlotServer(id, filename, plot)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{filename}{A string specifying a file name.}

\item{plot}{A plot object.}
}
\value{
The return value, if any, from executing the module server function
}
\description{
Create a plot download server.

\preformatted{
  downloadPlotServer("download_plot", filename = "tmp", data = data)
}
}
\seealso{
\code{\link[=downloadPlotUI]{downloadPlotUI()}}
}
