% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{downloadPlotUI}
\alias{downloadPlotUI}
\title{Plot Download Button UI}
\usage{
downloadPlotUI(id, label, right = FALSE, up = TRUE, animate = TRUE)
}
\arguments{
\item{id}{The name of the output slot that the downloadHandler is assigned to.}

\item{label}{The label that should appear on the button.}

\item{right}{A boolean specifying right or not.}

\item{up}{A boolean specifying up or not.}

\item{animate}{A boolean whether to use animation or not.}
}
\value{
A plot download button UI
}
\description{
Create a plot download button UI.

\preformatted{
  downloadPlotUI("download_plot", label = "download")
}
}
\seealso{
\code{\link[=downloadPlotServer]{downloadPlotServer()}}
}
