% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkboxGroup.R
\name{dynCheckboxGroupServer}
\alias{dynCheckboxGroupServer}
\title{Dynamic Checkbox Group Server}
\usage{
dynCheckboxGroupServer(id, data, column, selected = NULL, reverse = FALSE)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{data}{A reactive data}

\item{column}{A data column for unique list of values}

\item{selected}{The initially selected value.}

\item{reverse}{A boolean value whether to reverse the choices or not}
}
\value{
The return value, if any, from executing the module server function
}
\description{
Create a dynamic checkbox group server.
}
\details{
\preformatted{
## Example
dynCheckboxGroupServer("checkboxgroup", data, column, selected, reverse = FALSE)
}
}
\seealso{
\code{\link[=checkboxGroupUI]{checkboxGroupUI()}}
}
