% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{dynMultiServer}
\alias{dynMultiServer}
\title{Dynamic Multiselect Control Server}
\usage{
dynMultiServer(id, data, column, selected = character(0), reverse = FALSE)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{data}{A reactive data.}

\item{column}{A data column for unique list of values.}

\item{selected}{The initially selected value.}

\item{reverse}{A boolean value whether to reverse the choices or not}
}
\value{
The return value, if any, from executing the module server function
}
\description{
Create a dynamic multiselect control server.
}
\seealso{
\code{\link[=multiUI]{multiUI()}}
}
