% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric.R
\name{dynNumericServer}
\alias{dynNumericServer}
\title{Dynamic Numeric Input Control Server}
\usage{
dynNumericServer(id, data, column, value = 0)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{data}{A reactive data.}

\item{column}{A data column for calculating range of values.}

\item{value}{Initial value.}
}
\value{
The return value, if any, from executing the module server function
}
\description{
Create a dynamic numeric input control server.
}
\seealso{
\code{\link[=numericUI]{numericUI()}}
}
