% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectize.R
\name{dynSelectizeServer}
\alias{dynSelectizeServer}
\title{Dynamic Selectize Input Control Server}
\usage{
dynSelectizeServer(
  id,
  data,
  column,
  selected = NULL,
  reverse = FALSE,
  options = NULL,
  server = FALSE
)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{data}{A reactive data.}

\item{column}{A data column for unique list of values.}

\item{selected}{The initially selected value.}

\item{reverse}{A boolean value whether to reverse the choices or not}

\item{options}{A list of options. See the documentation of
\pkg{selectize.js}(\url{https://selectize.dev/docs/usage}) for possible options}

\item{server}{whether to store choices on the server side, and load the select options dynamically on searching,
instead of writing all choices into the page at once (i.e., only use the client-side version of \pkg{selectize.js})}
}
\value{
The return value, if any, from executing the module server function
}
\description{
Create a dynamic selectize input control server.
}
\seealso{
\code{\link[=selectizeUI]{selectizeUI()}}
}
