% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{dynSelectServer}
\alias{dynSelectServer}
\title{Dynamic Select Input Control Server}
\usage{
dynSelectServer(id, data, column, selected = NULL, reverse = FALSE)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{data}{A reactive data.}

\item{column}{A data column for unique list of values.}

\item{selected}{The initially selected value.}

\item{reverse}{A boolean value whether to reverse the choices or not}
}
\value{
The return value, if any, from executing the module server function
}
\description{
Create a dynamic select input control server.
}
\seealso{
\code{\link[=selectUI]{selectUI()}}
}
