% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slider.R
\name{dynSliderServer}
\alias{dynSliderServer}
\title{Dynamic Slider Input Control Server}
\usage{
dynSliderServer(id, data, column, value = NULL)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{data}{A reactive data.}

\item{column}{A data column for unique list of values.}

\item{value}{The initial value of the slider, either a number, a date
(class Date), or a date-time (class POSIXt). A length one vector will
create a regular slider; a length two vector will create a double-ended
range slider. Must lie between min and max of the \code{column}.}
}
\value{
The return value, if any, from executing the module server function
}
\description{
Create a dynamic slider input control server.
}
\seealso{
\code{\link[=sliderUI]{sliderUI()}}
}
