% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentalMeasuresModule.R
\name{ExperimentalMeasuresUI}
\alias{ExperimentalMeasuresUI}
\title{the UI for a very simple message box module}
\usage{
ExperimentalMeasuresUI(
  id,
  title,
  boxWidth = 300,
  boxHeight = 300,
  fontSize = 20,
  backgroundColor = "beige"
)
}
\arguments{
\item{id}{character string, the html document's widget id}

\item{title}{character string, title for the box}

\item{boxWidth}{integer pixels, 200 by default}

\item{boxHeight}{integer pixels, 30 by default}

\item{fontSize}{integer pixels, 20 by default}

\item{backgroundColor}{character string, uses standard CSS naming conventions, "beige" by default}
}
\description{
the UI for a very simple message box module
}
