% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{fileServer}
\alias{fileServer}
\title{File Control Server}
\usage{
fileServer(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}
}
\value{
The return value, if any, from executing the module server function
}
\description{
Create a file control server.
}
\details{
\preformatted{
## Example
fileServer("file_load")
}
}
\seealso{
\code{\link[=fileUI]{fileUI()}}
}
