% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{fileUI}
\alias{fileUI}
\title{File Control UI}
\usage{
fileUI(id, label = "Files")
}
\arguments{
\item{id}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}
}
\value{
A file control UI
}
\description{
Create a file control UI.

\preformatted{
## Example
fileUI("file_load", label = "Files")
}
}
\seealso{
\code{\link[=fileServer]{fileServer()}}
}
