% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter_dataUI}
\alias{filter_dataUI}
\title{UI part of the module filter_data}
\usage{
filter_dataUI(id)
}
\arguments{
\item{id}{\code{character} An id that will be used to create a namespace}
}
\value{
UI page
}
\description{
This function has to be set in the UI part of a shiny 
application. filter_data is a module used to filter a given data.table on 
chosen columns by the user. The server part returns a reactive value containing
the filtered data.table
}
\examples{
\dontrun{
# In UI :
filter_dataUI(id = "id")
# In Server
data <- reactiveValues(data = iris)
callModule(module = filter_data, id = "id", data = reactive(data))

## For a complete example, you can run the function \\link{run_example_app_filter_data}
run_example_app_filter_data()

## For another example which also uses the module \\link{show_data}, you can
## run \\link{run_example_app_filter_and_show_data}
optional_stats <- "all"
run_example_app_filter_and_show_data(optional_stats)
} 
}
