% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gen_reactable.R
\name{gen_reactable}
\alias{gen_reactable}
\title{Generate reactable from tibble}
\usage{
gen_reactable(
  .tbl,
  .col_names = NULL,
  .col_sortable = NULL,
  .selection = NULL,
  .searchable = FALSE
)
}
\arguments{
\item{.tbl}{(tbl_df) Data to display in table}

\item{.col_names}{(character) Vector of column names}

\item{.col_sortable}{(character) Vector of column names to be sortable}
}
\value{
(reactable) htmlwidget table
}
\description{
Generate reactable from tibble
}
