% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicators.R
\name{get_dt_num_dt_fac}
\alias{get_dt_num_dt_fac}
\title{preprocessing on data}
\usage{
get_dt_num_dt_fac(data, optional_stats)
}
\arguments{
\item{data}{\code{data.table} input data which will be preprocessed}

\item{optional_stats}{\code{character} optional statistics computed on data,
you can look at \link{show_data} for more information.}
}
\value{
list with two DT, one with statistics on numeric data and one with
statistics on factor data
}
\description{
preprocessing of input data, return two DT, one with
statistics onnumeric variables and another one with statistics on factor variables
}
\examples{
\dontrun{
get_dt_num_dt_fac(mtcars, optional_stats = c("min", "max", "boxplot", "density"))

get_dt_num_dt_fac(airquality, optional_stats = "all")

}

}
