% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_heatmap.R
\name{get_Heatmapdt}
\alias{get_Heatmapdt}
\title{Heatmap function with the DT package}
\usage{
get_Heatmapdt(data)
}
\arguments{
\item{data}{\code{data.table} data.table with two character columns}
}
\value{
DT in a heatmap form
}
\description{
This function takes data with two factor/character variables
and return a DT object in a heatmap form
}
\examples{
\dontrun{
titanic <- data.table(datasets::Titanic)
get_Heatmapdt(titanic[, list(Class, Age)])


}
}
