% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicators.R
\name{get_summary_data_dt}
\alias{get_summary_data_dt}
\title{Compute summary table(s)}
\usage{
get_summary_data_dt(
  data,
  optional_stats,
  nb_modal2show,
  show_warnings = FALSE,
  keep_dataframe = TRUE,
  keep_datatable = TRUE,
  session = NULL,
  message = "Calculation in progress..."
)
}
\arguments{
\item{data}{\code{data.frame}, \code{data.table} input data which will be preprocessed}

\item{optional_stats}{\code{character} optional statistics computed on data,
you can look at \link{summary_data} for more information.}

\item{nb_modal2show}{\code{integer} number of modalities to show for factor variables.}

\item{show_warnings}{\code{logical} Show warnings ? (example compute Min. on all NAs)}

\item{keep_dataframe}{\code{logical} data.frame in output ?}

\item{keep_datatable}{\code{logical} datatable (DT) in output ?}

\item{session}{the Shiny session object (from the server function of the Shiny app) if relevant. For progress bar.}

\item{message}{\code{character} Message of the progress computing indicators.}
}
\value{
list with two DT, one with statistics on numeric data and one with
statistics on factor data
}
\description{
preprocessing of input data, return one or more DT, one with
statistics on numeric variables, another one with statistics on factor variables
and one with date variables
}
\examples{
\dontrun{

get_summary_data_dt(mtcars, optional_stats = c("min", "max", "boxplot", "density"))

get_summary_data_dt(airquality, optional_stats = "all")

}

}
