% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggAnalyteLinePlotModule.R
\name{ggAnalyteLinePlotServer}
\alias{ggAnalyteLinePlotServer}
\title{the server for a DataTable shiny module}
\usage{
ggAnalyteLinePlotServer(input, output, session, plotTitle, data)
}
\arguments{
\item{input}{enviroment provide by shiny}

\item{output}{enviroment provide by shiny}

\item{session}{enviroment provide by shiny}

\item{plotTitle}{character string}

\item{data}{data.frame with area and time columns}
}
\description{
the server for a DataTable shiny module
}
