% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggAnalyteLinePlotModule.R
\name{ggAnalyteLinePlotUI}
\alias{ggAnalyteLinePlotUI}
\title{the UI for a ggAnalyteLinePlot shiny module}
\usage{
ggAnalyteLinePlotUI(id)
}
\arguments{
\item{id}{the html document's widget id}
}
\description{
the UI for a ggAnalyteLinePlot shiny module
}
