% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_from_connection.R
\name{give_info_data_sql}
\alias{give_info_data_sql}
\title{preprocessing on data}
\usage{
give_info_data_sql(con)
}
\arguments{
\item{con}{\code{SQLiteConnection} SQLite connection}
}
\description{
preprocessing of input data, return two DT, one with
statistics on numeric variables and another one with statistics on factor variables
}
