% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_heatmap.R
\name{heatmap}
\alias{heatmap}
\title{Make chart}
\usage{
heatmap(
  data,
  classes,
  labels = TRUE,
  cex = 10,
  main = "",
  xLabelsRotation = 45,
  colorby = "all",
  rownames,
  col = c("#FF0000", "#FFFFFF", "#0000FF")
)
}
\arguments{
\item{data}{: data.frame}

\item{classes}{: classes}

\item{labels}{: TRUE FALSE, display labels}

\item{cex}{: size of labels}

\item{main}{: Title of the heatmap}

\item{xLabelsRotation}{: rotation of xlabels}

\item{colorby}{: can be "all","row","col".}

\item{rownames}{: Rownames of the heatmap graphic}

\item{col}{: 3 col to use in colorRampPalette}
}
\value{
data.frame compound to original data.frame and associated constructor data.frame
}
\description{
Make chart
}
