% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap_controls_ui}
\alias{heatmap_controls_ui}
\title{A heatmap controls UI module}
\usage{
heatmap_controls_ui(
  id,
  label = "Expression Heatmap",
  transform_choice_names = c("Raw", "Mean Centred and Scaled", "log10", "Max Scaled"),
  transform_choices = c("raw", "mcs", "log10", "max"),
  cluster_choice_names = c("By Genes", "By Samples"),
  cluster_choices = c("rows", "columns")
)
}
\arguments{
\item{id}{id for the UI element. Must be matched when calling
the heatmap_controls_server function from callModule. [character]}

\item{label}{label for the UI control. [character]}

\item{transform_choice_names}{vector of names for the Transform radio buttons}

\item{transform_choices}{vector of choices for the Transform radio buttons}

\item{cluster_choice_names}{vector of names for the Cluster checkbox group}

\item{cluster_choices}{vector of choices for the Cluster checkbox group}
}
\value{
a tagList of UI elements
}
\description{
\code{heatmap_controls_ui} Controls for a heatmap plot
}
\examples{
ui <- fluidPage(
  fluidRow(
      wellPanel(heatmap_controls_ui("expr_heatmap_controls", label = "Expression Heatmap"))
  )
)

}
