% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap_server}
\alias{heatmap_server}
\title{A heatmap server module}
\usage{
heatmap_server(input, output, session, data, heatmap_selections)
}
\arguments{
\item{data}{The data used to create the heatmap}

\item{heatmap_selections}{The selections made by the user in the controls UI section}
}
\value{
a matrix, transformed and clustered
}
\description{
\code{heatmap_server} The server side code to produce a heatmap plot.
It produces the heatmap and returns the data used to create it.
}
\examples{

server <- function(input, output, session) {
  heatmap_data <- callModule(heatmap_server, "expr_heatmap", heatmap_selections)
}
}
