% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap_ui}
\alias{heatmap_ui}
\title{A heatmap UI module}
\usage{
heatmap_ui(id)
}
\arguments{
\item{id}{id for the UI element. Must be matched when calling
the heatmap_server function from callModule. [character]}
}
\value{
a plotOutput object
}
\description{
\code{heatmap_ui} The output of a heatmap plot
}
\examples{
ui <- fluidPage(
  fluidRow(
      heatmap_ui("expr_heatmap"))
  )
)

}
