% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmapApp}
\alias{heatmapApp}
\title{A test shiny app for the heatmap module}
\usage{
heatmapApp()
}
\value{
a [shiny::shinyApp()] object.
}
\description{
`heatmapApp()` creates a small test app for testing the [heatmapOutput()] and
[heatmapServer()] functions. It uses a subset of the package datasets `counts`,
`sampleInfo` and `gene_metadata` and create a [ggplot2::ggplot()] heatmap object.
It also has transform radio buttons to test using that information to change
the colour palette.
}
\examples{

heatmapApp()

}
