% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmapInput}
\alias{heatmapInput}
\title{Create UI components to display heatmap of a matrix of values}
\usage{
heatmapInput(id)
}
\arguments{
\item{id}{namespace id for the UI components. Must match the id provided to the
[heatmapServer()] function.}
}
\value{
a [htmltools::tagList()] containing a [shiny::radioButtons] object
}
\description{
`heatmapInput()` produces a control for changing the colour scale of the heatmap
}
\examples{

heatmapInput("rnaseqData")

}
