% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmapServer}
\alias{heatmapServer}
\title{Server function to create a heatmap of a count matrix}
\usage{
heatmapServer(
  id,
  counts = NULL,
  sample_info = NULL,
  gene_metadata = NULL,
  transform = NULL,
  debug = NULL
)
}
\arguments{
\item{id}{namespace id for the UI components. Must match the id provided to the
[heatmapInput()] function.}

\item{counts}{a reactive counts object. Should contain only numeric columns}

\item{sample_info}{a reactive object. Represents the samples and associated
metadata}

\item{gene_metadata}{a reactive object. Contains the metadata for the genes
present in the counts object.}

\item{transform}{reactive that contains the contents of input$transform from
the transform module}

\item{debug}{Turn on debugging message statements}
}
\value{
a [shiny::reactive()] object which is the heatmaped counts
}
\description{
`heatmapServer()` creates a [ggplot2] plot object using [biovisr::matrix_heatmap]
from the supplied count matrix, sample info and gene metadata
}
\examples{

heatmapServer("rnaseqData", counts = reactive(rnaseqtools::counts[1:10,1:5]))

}
