% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iframeSearchModule.R
\name{iframeSearchServer}
\alias{iframeSearchServer}
\title{the server for a DataTable shiny module}
\usage{
iframeSearchServer(input, output, session, website, geneSymbol)
}
\arguments{
\item{input}{enviroment provide by shiny}

\item{output}{enviroment provide by shiny}

\item{session}{enviroment provide by shiny}

\item{website}{character string, one of genecards, pubmed, homologene, kegg, dbsnp, google}

\item{searchTerm}{character string, eg a geneSymbol, an rsid, ...}
}
\description{
the server for a DataTable shiny module
}
