% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iframeSearchModule.R
\name{iframeSearchUI}
\alias{iframeSearchUI}
\title{the UI for a shiny module to display (in an iframe) a search of a 3rd party website (pubmed, google, ...)}
\usage{
iframeSearchUI(id, title)
}
\arguments{
\item{id}{the html document's widget id}
}
\description{
the UI for a shiny module to display (in an iframe) a search of a 3rd party website (pubmed, google, ...)
}
