% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igvModule.R
\name{igvServer}
\alias{igvServer}
\title{the server for an igv.js shiny module}
\usage{
igvServer(
  id,
  input,
  output,
  session,
  genome,
  locus,
  width = 800,
  height = 800,
  geneModelDisplayMode = "COLLAPSED",
  geneModelTrackHeight = 200
)
}
\arguments{
\item{input}{enviroment provide by shiny}

\item{output}{enviroment provide by shiny}

\item{session}{enviroment provide by shiny}

\item{genome}{character string, eg "hg38"}

\item{locus}{character string, eg "APOE" or "chr13:1000000-1001050"}

\item{geneModelDisplayMode}{character string, COLLAPSED, SQUISHED, EXPANDED}

\item{geneModeTrackHeight}{integer, eg 200}
}
\description{
the server for an igv.js shiny module
}
