% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadServer}
\alias{loadServer}
\title{Load Server}
\usage{
loadServer(id, filename, loadfun = "readRDS")
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{filename}{A string specifying the file name.}

\item{loadfun}{A string specifying the function name to use when loading data.}
}
\value{
The return value, if any, from executing the module server function
}
\description{
Create a load server.

\preformatted{
  loadServer("load_id", filename = "tmp.rds", loadfun = "readRDS")
}
}
\seealso{
\code{\link[=loadUI]{loadUI()}}
}
