% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadUI}
\alias{loadUI}
\title{Load Button UI}
\usage{
loadUI(id, label = "Load")
}
\arguments{
\item{id}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}
}
\value{
A load button UI
}
\description{
Create a load button UI.

\preformatted{
  loadUI("load", label = "Load")
}
}
\seealso{
\code{\link[=loadServer]{loadServer()}}
}
