% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logScale.R
\name{logScaleUI}
\alias{logScaleUI}
\title{Log scale Control UI}
\usage{
logScaleUI(id, label = "Log scale", choices = NULL, selected = NULL)
}
\arguments{
\item{id}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}

\item{choices}{Default NULL, list of values to select from.}

\item{selected}{Default NULL, the initially selected value.}
}
\value{
A log scale control UI
}
\description{
Create a log scale control UI.

\preformatted{
## Example
logScaleUI("logscale")
}
}
\seealso{
\code{\link[=logScaleServer]{logScaleServer()}}
}
