% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{mapModule}
\alias{mapModule}
\title{Server function of the mapmodule}
\usage{
mapModule(input, output, session)
}
\arguments{
\item{NA}{No parameters yet}
}
\value{
NA It plots a map
}
\description{
Creates a leaflet map. More features to come
}
\examples{
to come
}

