% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{mapModuleUI}
\alias{mapModuleUI}
\title{UI function of the mapmodule}
\usage{
mapModuleUI(id)
}
\arguments{
\item{id}{The is the id that needs to be specified in both the UI and server function of the shiny app}
}
\value{
NA It plots a map
}
\description{
Creates a leaflet map. More features to come
}
\examples{
to come
}

