% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{max_scale}
\alias{max_scale}
\title{Scale count according to the maximum value}
\usage{
max_scale(counts)
}
\arguments{
\item{counts}{tibble of counts}
}
\value{
a tibble of max sclaed counts
}
\description{
Scale count according to the maximum value
}
\examples{

max_scale(rnaseqtools::counts)

}
