% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messageBoxModule.R
\name{messageBoxServer}
\alias{messageBoxServer}
\title{the server for a MessageBox shiny module}
\usage{
messageBoxServer(id, input, output, session, newContent)
}
\arguments{
\item{input}{enviroment provide by shiny}

\item{output}{enviroment provide by shiny}

\item{session}{enviroment provide by shiny}

\item{newContent}{character string, what to display}
}
\description{
the server for a MessageBox shiny module
}
