% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messageBoxModule.R
\name{messageBoxUI}
\alias{messageBoxUI}
\title{the UI for a very simple message box module}
\usage{
messageBoxUI(id, title, titleSpan = 1, boxSpan = 11)
}
\arguments{
\item{id}{character string, the html document's widget id}

\item{title}{character string, title for the box}

\item{titleSpan}{integer, width in bootstrap units (max totaling 12)}

\item{boxSpan}{integer, width in bootstrap units (max totaling 12)}
}
\description{
the UI for a very simple message box module
}
