% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_data.R
\name{monitoring_data}
\alias{monitoring_data}
\alias{monitoring_data_UI}
\title{Shiny module to display monitoring indicators on a data.table.}
\usage{
monitoring_data(
  input,
  output,
  session,
  data,
  col_id = NULL,
  keep_id = NULL,
  col_obs,
  col_fit,
  col_date = NULL,
  tz = "UTC",
  indicators = c("rmse", "mae", "mape", "mape_e"),
  labels = list(progress_data = "Processing data", no_data = "No data", no_choice =
    "None", idc_title = "Distribution of indicators", idc_aggr = "Aggregation column",
    idv_opt_discretiser = "Discretise ?", idv_choice_quantiles = "Quantiles choice",
    idc_button = "Display graphic", idc_progress = "Processing indicators",
    idc_plot_title = "Model's indicators", err_title = "Distribution of errors", err_aggr
    = "Aggregation column", err_type = "Error type choice", err_button =
    "Display graphic", err_progress = "Processing boxplots",      err_ylab = "Values",
    error_xlab = "error : ", tree_title = "Decision tree", tree_y = "Y", tree_x = "X",
    tree_minsplit = "Minsplit", tree_cp = "Complexity (cp)", tree_maj_cp = "Update cp",
    tree_maj_params = "Update parameters", tree_cp_modal_titre = "Update cp",
    tree_cp_modal_min = "Slider min", tree_cp_modal_max = "Slider max",
    tree_cp_modal_step = "Slider step", tree_cp_modal_bouton = "Validate update",
    warning_var = "Select at least one explanatory variable.")
)

monitoring_data_UI(id)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny input}

\item{session}{shiny input}

\item{data}{\code{data.frame} / \code{data.table}. Table on which indicators will be computed.}

\item{col_id}{\code{character} (NULL). Column name for id values.}

\item{keep_id}{\code{character} (NULL). id values to keep in col_id.}

\item{col_obs}{\code{character}. Column name for observed values.}

\item{col_fit}{\code{character}. Column name for fitted values.}

\item{col_date}{\code{character} (NULL). Column name for date values.}

\item{tz}{\code{character}. Timezone of new columns. Default to "UTC"}

\item{indicators}{\code{characters}. Indicators to be computed, amongst: ("rmse", "mae", "mape", "mape_star").}

\item{labels}{\code{character}. Labels to modify displayed texts. See default in examples.}

\item{id}{\code{character}. shiny id to allow multiple instanciation.}
}
\value{
shiny module.
}
\description{
Shiny module to display monitoring indicators on a data.table.
}
\examples{
\dontrun{\donttest{

library(data.table)

data <- data.table(obs = runif(100, 1, 10))
data[, fit := obs + rnorm(100, 0, 10)]
data[, date := seq(as.POSIXct("2019-10-07 00:00:00 UTC", tz = "UTC"),
                   as.POSIXct("2019-10-11 03:00:00 UTC", tz = "UTC"), by = 60*60)]
data[, by_quali := factor(sample(rep(1:10, 10)))]
data[, by_quanti := runif(100, 1, 20)]

col_obs <- "obs"
col_fit <- "fit"
col_date = "date"
indicators <- c("rmse", "mae", "mape", "mape_e")

ui <- monitoring_data_UI(id = "my_id")

server <- function(input, output, session) {
  callModule(module = monitoring_data, id = "my_id", 
             data = reactive(data),
             col_obs = col_obs,
             col_fit = col_fit,
             col_date = col_date,
             indicators = indicators
  )
}
shiny::shinyApp(ui = ui, server = server)

}}

}
