% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_data.R
\name{monitoring_data_UI}
\alias{monitoring_data_UI}
\title{Shiny module ui-like fun to display indicators on a data.table.}
\usage{
monitoring_data_UI(id)
}
\arguments{
\item{id}{\code{character}. shiny id to allow multiple instanciation.}
}
\value{
shiny module.
}
\description{
Shiny module ui-like fun to display indicators on a data.table.
}
\examples{
\dontrun{\donttest{

library(data.table)

data <- data.table(obs = runif(100, 1, 10))
data[, fit := obs + rnorm(100, 0, 10)]
data[, date := seq(as.POSIXct("2019-10-07 00:00:00 UTC", tz = "UTC"),
                   as.POSIXct("2019-10-11 03:00:00 UTC", tz = "UTC"), by = 60*60)]
data[, by_quali := factor(sample(rep(1:10, 10)))]
data[, by_quanti := runif(100, 1, 20)]

col_obs <- "obs"
col_fit <- "fit"
col_date = "date"
indicators <- c("rmse", "mae", "mape", "mape_e")

ui <- monitoring_data_UI("my_id")
server <- function(input, output, session) {
  callModule(monitoring_data, "my_id", data = reactive(data), 
             col_obs = col_obs, 
             col_fit = col_fit, 
             col_date = col_date,
             indicators = indicators)
}
shiny::shinyApp(ui = ui, server = server)

}}
}
