% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{multiServer}
\alias{multiServer}
\title{Multiselect Control Server}
\usage{
multiServer(id, choices = character(0), selected = character(0))
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{choices}{List of values to select from.}

\item{selected}{The initially selected value.}
}
\value{
The return value, if any, from executing the module server function
}
\description{
Create a multiselect control server.
}
\seealso{
\code{\link[=multiUI]{multiUI()}}
}
