% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{multiUI}
\alias{multiUI}
\title{Multiselect Control UI}
\usage{
multiUI(
  id,
  label = "",
  choices = character(0),
  selected = character(0),
  options = NULL,
  width = NULL
)
}
\arguments{
\item{id}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}

\item{choices}{List of values to select from.}

\item{selected}{The initially selected value.}

\item{options}{List of options passed to multi (enable_search = FALSE for disabling the search bar for example).}

\item{width}{width The width of the input, e.g. \verb{400px}, or \verb{100\%}.}
}
\value{
A multiselect control UI.
}
\description{
Create a multiselect control UI.
}
\seealso{
\code{\link[=multiServer]{multiServer()}}, \code{\link[=dynMultiServer]{dynMultiServer()}}
}
