% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric.R
\name{numericServer}
\alias{numericServer}
\title{Numeric Input Control Server}
\usage{
numericServer(id, value = 0, min = 0, max = 1, step = 1)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{value}{Initial value.}

\item{min}{Minimum allowed value}

\item{max}{Maximum allowed value}

\item{step}{Interval to use when stepping between min and max}
}
\value{
The return value, if any, from executing the module server function
}
\description{
Create a numeric input control server.
}
\seealso{
\code{\link[=numericUI]{numericUI()}}
}
