% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric.R
\name{numericUI}
\alias{numericUI}
\title{Numeric Input Control UI}
\usage{
numericUI(id, label = "", value = 0, min = 0, max = 1, step = 1)
}
\arguments{
\item{id}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}

\item{value}{Initial value.}

\item{min}{Minimum allowed value}

\item{max}{Maximum allowed value}

\item{step}{Interval to use when stepping between min and max}
}
\value{
A numeric input control UI
}
\description{
Create a numeric input control UI.
}
\seealso{
\code{\link[=numericServer]{numericServer()}}, \code{\link[=dynNumericServer]{dynNumericServer()}}
}
