% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericinput_UI.R
\name{numericValueUI}
\alias{numericValueUI}
\title{Wrapper around shiny::numericInput.}
\usage{
numericValueUI(id,
  label = paste("Please provide a label argument to the numericInputUI",
  " with id = ", id, ".", sep = ""), value = 0, min = value - 3 * value,
  max = value + 3 * value, step = (max - min)/100, width = "10\%",
  custom_css = "")
}
\arguments{
\item{id}{A unique string id for the numericInput. If your code is modularised,
remember to wrap the id with the relevant session namespace function, e.g.
numericValueUI(id = session$ns("someinput")).}

\item{label}{A label for the input.}

\item{value}{The default value.}

\item{min}{The smallest allowed value.}

\item{max}{The largest allowed value.}

\item{step}{The step size when incrementing.}

\item{width}{The width of the shiny::numericValue UI element. Must be valid
css (i.e. "10%" or "150px" will work).}

\item{custom_css}{The numericInput will be wrapped in a div() styled with
the css code provided as a string with the custom_css input. For example,
when creating multiple instances of the same type of UI element, one could
provide custom_css = "display: inline-block;" to display the different
elements inline.}
}
\description{
Wrapper around shiny::numericInput.
}
