% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericinput_UI.R
\name{numericValueUIs}
\alias{numericValueUIs}
\title{Creates a tagList of numericInputUI instances.}
\usage{
numericValueUIs(ids, labels = NULL, values = NULL, minvals = NULL,
  maxvals = NULL, steps = NULL, widths = NULL, custom_css = NULL)
}
\arguments{
\item{ids}{A unique string id for the numericInput. If your code is modularised,
remember to wrap the id with the relevant session namespace function, e.g.
numericValueUI(id = session$ns("someinput")).}

\item{labels}{A label for the input.}

\item{values}{The default values.}

\item{minvals}{The smallest allowed values.}

\item{maxvals}{The largest allowed values.}

\item{steps}{The step sizes when incrementing.}

\item{widths}{The widths of the shiny::numericValue UI element. Must be valid
css (i.e. "10%" or "150px" will work).}

\item{custom_css}{The numericInput will each be wrapped in da iv() styled with
the css code provided as a string with the custom_css input. For example,
when creating multiple instances of the same type of UI element, one could
provide custom_css = "display: inline-block;" to display the different
elements inline.}
}
\description{
Creates a tagList of numericInputUI instances.
}
