% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_factors.R
\name{order_variable_rb_server}
\alias{order_variable_rb_server}
\title{A server function to get the value selected}
\usage{
order_variable_rb_server(input, output, session)
}
\arguments{
\item{input, output, session}{standard \code{shiny} boilerplate}
}
\value{
list with following components
\describe{
  \item{order_by}{reactive character indicating selection}
}

server <- function(input, output, session) {
  x_ordering <- callModule(order_variable_rb_server, "order_x")
}
}
\description{
\code{order_variable_rb_server} returns a list of selections
made in the heatmap control UI module. This function should
be used with callModule.
}
