% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_factors.R
\name{order_variable_rb_ui}
\alias{order_variable_rb_ui}
\title{A UI module for choosing how variables are ordered}
\usage{
order_variable_rb_ui(
  id,
  label = "X variable",
  choice_names = c("Alphabetical", "Numeric", "Order of Appearance", "By Frequency"),
  choices = c("alpha", "numeric", "appearance", "by_freq")
)
}
\arguments{
\item{id}{id for the UI element. Must be matched when calling
the server function from callModule. [character]}

\item{label}{label for the UI control. [character]}

\item{choice_names}{vector of names for the radio buttons}

\item{choices}{vector of choices for the radio buttons}
}
\value{
a RadioButtons UI element
}
\description{
\code{order_variable_rb_ui} Produces a UI component for selecting how
to order a variable
}
\examples{
ui <- fluidPage(
  fluidRow(
      wellPanel(order_variable_rb_ui("order_x", label = "Order X Variable",
          choice_names = c('Alphabetical', 'Numeric', 'Order of Appearance', 'By Frequency'),
          choices= c('alpha', 'numeric', 'appearance', 'by_freq')))
  )
)

}
