% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_factors.R
\name{order_variable_server}
\alias{order_variable_server}
\title{A server function to order a factor}
\usage{
order_variable_server(input, output, session, fct_to_order, ordering)
}
\arguments{
\item{input, output, session}{standard \code{shiny} boilerplate}

\item{ordering}{reactive list containing order_by}

\item{data}{reactive data source}

\item{xvar}{character}
}
\value{
The data with the reordered factor

server <- function(input, output, session) {
  x_ordering <- callModule(order_variable_server, "order_x")
}
}
\description{
\code{order_variable_server} returns a list of selections
made in the heatmap control UI module. This function should
be used with callModule.
}
