% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plots.R
\name{plotExploratory}
\alias{plotExploratory}
\title{plot boxplot, timeseries and barplot}
\usage{
plotExploratory(data, type, palette_ggplot = "RdYlGn",
  aggregation = NULL, js = TRUE)
}
\arguments{
\item{data}{\code{data.table} data.table with one or two columns, one numeric and 
another one, character, factor or date}

\item{type}{\code{character} type of the graph, default is line, possible 
values are "boxplot", "barplot" and "timeseries"}

\item{palette_ggplot}{\code{character} palette you want to use for your graphics
(one of the palettes from the package RColorBrew)}

\item{aggregation}{\code{character} how you want to do the aggregation 
(if timeseries), possible values are "Average", "Sum", "Low" and "High"}

\item{js}{\code{logical} TRUE if you want dynamic graph}
}
\description{
This function takes a data.table with one numeric
variable and one character, factor or date variable in input and return 
a graph of the type chosen by the user, it can do interactive or fix graph.
}
\examples{
\dontrun{
iris <- data.table::copy(data.table(datasets::iris))
flights <- data.table(nycflights13::flights)

plotExploratory(iris[, list(Petal.Width, Species)], type = "barplot")
plotExploratory(iris[, list(Sepal.Width)], type = "boxplot")
plotExploratory(flights[, list(time_hour, distance)], type = "timeseries",
aggregation = "Average", js = F)


}

}
