% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plots.R
\name{plotScatterplot}
\alias{plotScatterplot}
\title{plot histogram and scatterplot}
\usage{
plotScatterplot(data, type = "line", js = TRUE, id = NULL)
}
\arguments{
\item{data}{\code{data.table} data.table with one or two numeric columns}

\item{type}{\code{character} type of the graph, default is line, possible 
values are "line", "point" and "hist"}

\item{js}{\code{logical} if TRUE, the graph is interactive (default TRUE)}

\item{id}{\code{character} if you use an id column, default NULL}
}
\description{
This function takes a data.table with one or two numeric
variables in input and return a graph of the type chosen by the user,
it can do interactive or fix graph.
}
\examples{
\dontrun{
mtcars <- data.table::copy(data.table(datasets::mtcars))

plotScatterplot(mtcars[, list(mpg, cyl)], type = "line", js = T)
plotScatterplot(mtcars[, list(mpg, drat)], type = "point", js = T)
plotScatterplot(mtcars[, list(drat)], type = "hist", js = T)
plotScatterplot(mtcars[, list(drat, wt)], type = "point", js = F)
}

}
