% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_apps_tests.R
\name{run_app_show_data}
\alias{run_app_show_data}
\title{Run a shiny example application to visualise statistics}
\usage{
run_app_show_data(optional_stats)
}
\arguments{
\item{optional_stats}{\code{character} vector of the optional statistics you
want to be computed on your numeric data, default is "all". "pct_zero", 
"pct_NA", "mean", "median", "sd" are always computed, possible values are 
"min", "max", "var", "ecart_interquartile",
"mode_max", "kurtosis", "skewness", "boxplot", "density".}
}
\description{
Run a shiny application calling the module show_data, displaying two tables 
with descriptive statistics, one for numeric variables and one for factor variables.
}
\examples{

\dontrun{

optional_stats <- "all"
run_app_show_data(optional_stats)

optional_stats <- c("min", "max", "var", "boxplot", "density")
run_app_show_data(optional_stats)
}

}
