% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_apps_tests.R
\name{run_example_app_filter_and_show_data}
\alias{run_example_app_filter_and_show_data}
\title{Run a shiny example application to filter on data and
 visualise statistics on the filtered data}
\usage{
run_example_app_filter_and_show_data(optional_stats = "all",
  nb_modal2show = 5)
}
\arguments{
\item{optional_stats}{\code{character} vector of the optional statistics you
want to be computed on your numeric data, default is "all". "pct_zero", 
"pct_NA", "mean", "median", "sd" are always computed, possible values are 
"min", "max", "var", "ecart_interquartile",
"mode_max", "kurtosis", "skewness", "boxplot", "density".}

\item{nb_modal2show}{\code{integer} number of modalities to show 
for factor variables.}
}
\description{
Run a shiny application calling the modules show_data and filter_data, 
allowing the user to use filters on the different variables of his data 
and then displaying two tables with descriptive statistics, 
one for numeric variables and one for factor variables.
}
\examples{
\dontrun{
run_example_app_filter_and_show_data(optional_stats)

optional_stats <- c("min", "max", "var", "boxplot", "density")
nb_modal2show <- 5
run_example_app_filter_and_show_data(optional_stats)
}

}
