% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_apps_tests.R
\name{run_example_app_filter_data}
\alias{run_example_app_filter_data}
\title{Run a shiny example application to filter data and download the filtered
data}
\usage{
run_example_app_filter_data()
}
\description{
Run a shiny application calling the module filter_data, allowing the user
to use filters on the different variables of his data
}
\examples{

\dontrun{

run_example_app_filter_data()
}

}
