% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_apps_tests.R
\name{run_example_app_monitoring_data}
\alias{run_example_app_monitoring_data}
\title{Run a shiny example application to display monitoring indicators}
\usage{
run_example_app_monitoring_data()
}
\description{
Run a shiny application calling the module monitoring_data
}
\examples{

\dontrun{

run_example_app_monitoring_data()
}

}
