% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_apps_tests.R
\name{run_example_app_show_data}
\alias{run_example_app_show_data}
\title{Run a shiny example application to visualise statistics}
\usage{
run_example_app_show_data(optional_stats = "all", nb_modal2show = 5)
}
\arguments{
\item{optional_stats}{\code{character} vector of the optional statistics you
want to be computed on your numeric data, default is "all". "pct_zero", 
"pct_NA", "mean", "median", "sd" are always computed, possible values are 
"min", "max", "var", "interquartile_range",
"mode_max", "kurtosis", "skewness", "boxplot", "density".}

\item{nb_modal2show}{\code{integer} number of modalities to show for factor variables.}
}
\description{
Run a shiny application calling the module show_data, displaying three tables 
with descriptive statistics, one for numeric variables, one for dates and one for factor variables.
}
\examples{

\dontrun{

optional_stats <- "all"
nb_modal2show <- 5
run_example_app_show_data(optional_stats, nb_modal2show)

optional_stats <- c("min", "max", "var", "boxplot", "density")
run_example_app_show_data(optional_stats, nb_modal2show)
}

}
