% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_apps_tests.R
\name{run_example_app_visualize_data}
\alias{run_example_app_visualize_data}
\title{Run a shiny example application to visualize graphics on data}
\usage{
run_example_app_visualize_data()
}
\description{
Run a shiny application calling the module visualize_data, 
allowing the user to visualize his data with diffrent type of graphics
}
\examples{
\dontrun{

run_example_app_visualize_data()
}

}
