% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scalesUI}
\alias{scalesUI}
\title{Scales Control UI}
\usage{
scalesUI(id, label = "Scales", choices = NULL, selected = NULL)
}
\arguments{
\item{id}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}

\item{choices}{Default NULL, list of values to select from.}

\item{selected}{Default NULL, the initially selected value.}
}
\value{
A scale control UI
}
\description{
Create a scales control UI.

\preformatted{
## Example
scalesUI("scales")
}
}
\seealso{
\code{\link[=scalesServer]{scalesServer()}}
}
