% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_variables.R
\name{select_vars_pulldown_server}
\alias{select_vars_pulldown_server}
\title{A server function to get the value selected}
\usage{
select_vars_pulldown_server(input, output, session)
}
\arguments{
\item{input, output, session}{standard \code{shiny} boilerplate}
}
\value{
list with following components
\describe{
  \item{selected}{reactive character indicating choice selected}
}

server <- function(input, output, session) {
  xvar <- callModule(select_vars_pulldown_server, "xvar")
}
}
\description{
\code{select_vars_pulldown_server} returns a list of selections
made in the select_vars UI module. This function should
be used with callModule.
}
