% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_variables.R
\name{select_vars_pulldown_ui}
\alias{select_vars_pulldown_ui}
\title{A UI module for selecting variables (pull-down)}
\usage{
select_vars_pulldown_ui(id, label = "X variable", choices = NULL)
}
\arguments{
\item{id}{id for the UI element. Must be matched when calling
the server function from callModule. [character]}

\item{label}{label for the UI control. [character]}

\item{choices}{vector of choices for the pulldown list}
}
\value{
a selectInput UI element
}
\description{
\code{select_vars_pulldown_ui} Produces a UI component for selecting variables
from a list. The default option is always the first item in the list
}
\examples{
ui <- fluidPage(
  fluidRow(
      wellPanel(select_vars_pulldown_ui("xvar", label = "X Variable",
          choices = c('class', 'cyl', 'cty'))
  )
)

}
