% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectinputUI.R
\name{selectInputUIs}
\alias{selectInputUIs}
\title{Create tagList of selectInputUI instances.}
\usage{
selectInputUIs(ids, labels = NULL, choices = NULL, selectize = NULL,
  widths = NULL, custom_css = NULL)
}
\arguments{
\item{ids}{Vector of ids}

\item{labels}{Vector of labels for the dropdown menus}

\item{choices}{List of choice vectors for the dropdown menus. List elements
can be named; if so, they are grouped in the dropdown menu.}

\item{selectize}{Vector of booleans indiciating if dropdown menus should be
selectized.}

\item{widths}{Vector of valid css widths.}
}
\description{
Create tagList of selectInputUI instances.
}
