% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectize.R
\name{selectizeServer}
\alias{selectizeServer}
\title{Selectize Input Control Server}
\usage{
selectizeServer(id, choices, selected = NULL, options = NULL)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{choices}{Default NULL, list of values to select from.}

\item{selected}{Default NULL, the initially selected value.}

\item{options}{A list of options. See the documentation of
\pkg{selectize.js}(\url{https://selectize.dev/docs/usage}) for possible options}
}
\value{
A selectize input control server
}
\description{
Create a numeric input control server.
}
\seealso{
\code{\link[=selectizeUI]{selectizeUI()}}
}
