% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectize.R
\name{selectizeUI}
\alias{selectizeUI}
\title{Selectize Input Control UI}
\usage{
selectizeUI(
  id,
  label = "Select",
  choices = NULL,
  selected = NULL,
  multiple = FALSE,
  options = NULL
)
}
\arguments{
\item{id}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}

\item{choices}{Default NULL, list of values to select from.}

\item{selected}{Default NULL, the initially selected value.}

\item{multiple}{Is selection of multiple items allowed?}

\item{options}{A list of options. See the documentation of
\pkg{selectize.js}(\url{https://selectize.dev/docs/usage}) for possible options}
}
\value{
A selectize input control UI
}
\description{
Create selectize input control UI.
}
\seealso{
\code{\link[=selectizeServer]{selectizeServer()}}, \code{\link[=dynSelectizeServer]{dynSelectizeServer()}}
}
