% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{selectServer}
\alias{selectServer}
\title{Select Input Control Server}
\usage{
selectServer(id, choices, selected = NULL, selectize = TRUE)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{choices}{Default NULL, list of values to select from.}

\item{selected}{Default NULL, the initially selected value.}

\item{selectize}{Whether to use \pkg{selectize.js} or not.}
}
\value{
A select input control server
}
\description{
Create a numeric input control server.
}
\seealso{
\code{\link[=selectUI]{selectUI()}}
}
