% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{selectUI}
\alias{selectUI}
\title{Select Input Control UI}
\usage{
selectUI(
  id,
  label = "Select",
  choices = NULL,
  selected = NULL,
  multiple = FALSE,
  selectize = TRUE
)
}
\arguments{
\item{id}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}

\item{choices}{Default NULL, list of values to select from.}

\item{selected}{Default NULL, the initially selected value.}

\item{multiple}{Is selection of multiple items allowed?}

\item{selectize}{Whether to use \pkg{selectize.js} or not.}
}
\value{
A select input control UI
}
\description{
Create select input control UI.
}
\seealso{
\code{\link[=selectServer]{selectServer()}}, \code{\link[=dynSelectServer]{dynSelectServer()}}
}
