% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectVar.R
\name{SelectVarPulldownApp}
\alias{SelectVarPulldownApp}
\title{A test shiny app for the select_vars module}
\usage{
SelectVarPulldownApp(debug = FALSE)
}
\value{
a [shiny::shinyApp()] object
}
\description{
`SelectVarPulldownApp()` creates a small test app for testing the
[SelectVarPulldownUI()] and [SelectVarPulldownServer()] functions.
}
\examples{
SelectVarPulldownApp()

}
