% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectVar.R
\name{SelectVarPulldownServer}
\alias{SelectVarPulldownServer}
\title{A server function to get the value selected}
\usage{
SelectVarPulldownServer(id, data = NULL, debug = FALSE)
}
\arguments{
\item{input, output, session}{standard \code{shiny} boilerplate}
}
\value{
list with following components
\describe{
  \item{selected}{reactive character indicating choice selected}
}

selected_x <- SelectVarPulldownServer("xvar")
}
\description{
\code{SelectVarPulldownServer} returns a list of selections
made in the select_vars UI module. This function should
be used with callModule.
}
