% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectVar.R
\name{SelectVarPulldownUI}
\alias{SelectVarPulldownUI}
\title{A UI module for selecting variables (pull-down)}
\usage{
SelectVarPulldownUI(id, label = "X variable", data = NULL)
}
\arguments{
\item{id}{id for the UI element. Must be matched when calling
the server function. [character]}

\item{label}{label for the UI control. [character]}

\item{data}{data.frame to take the variables list from}
}
\value{
a varSelectizeInput UI element
}
\description{
\code{SelectVarPulldownUI} Produces a UI component for selecting variables
from a list. The default option is always the first item in the list
}
\examples{
ui <- fluidPage(
  fluidRow(
      wellPanel(SelectVarPulldownUI("xvar", label = "X Variable",
          choices = c('class', 'cyl', 'cty'))
    )
  )
)

}
