% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinymodules-package.R
\docType{package}
\name{shinymodules-package}
\alias{shinymodules}
\alias{shinymodules-package}
\title{shinymodules: A simple shiny modules}
\description{
Simple ui and server modules to avoid a complicated shiny app.
}
\details{
xxxUI functions are for an ui function
xxxServer functions are for a server function
}
\author{
\strong{Maintainer}: Seokhoon Joo \email{seokhoonj@gmail.com}

}
