% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_data.R
\name{show_data}
\alias{show_data}
\title{server part of the module show_data}
\usage{
show_data(input, output, session, data = NULL, optional_stats = "all",
  nb_modal2show = 3)
}
\arguments{
\item{input}{Not a real parameter, should not be set manually. 
Done by callModule automatically.}

\item{output}{Not a real parameter, should not be set manually. 
Done by callModule automatically.}

\item{session}{Not a real parameter, should not be set manually. 
Done by callModule automatically.}

\item{data}{\code{reactivevalues} reactive data.table}

\item{optional_stats}{\code{character} optional statistics computed on numeric
data, default is "all". "pct_zero", "pct_NA", "mean", "median", "sd" are
always computed, possible values are "min", "max", "var", "ecart_interquartile",
"mode_max", "kurtosis", "skewness", "boxplot", "density".}
}
\value{
Server logic
}
\description{
This function has to be set in the Server part of a shiny application
It returns two table, one with statistics on numeric data and one with 
statistics on factor data
}
\examples{
\dontrun{
# In UI :
show_dataUI(id = "id")
# In Server
callModule(module = show_data, id = "id", variable = reactive(data))

## For a complete example, you can run the function \\link{run_app_show_data}

optional_stats <- "all"
run_app_show_data(optional_stats)

}

}
