% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_data.R
\name{show_dataUI}
\alias{show_dataUI}
\alias{show_data}
\title{UI part of the module show_data}
\usage{
show_dataUI(id)

show_data(input, output, session, data = NULL, optional_stats = "all",
  nb_modal2show = 3)
}
\arguments{
\item{id}{\code{character} An id that will be used to create a namespace
It returns two table, one with statistics on numeric data and one with 
statistics on factor data}

\item{input}{Not a real parameter, should not be set manually. 
Done by callModule automatically.}

\item{output}{Not a real parameter, should not be set manually. 
Done by callModule automatically.}

\item{session}{Not a real parameter, should not be set manually. 
Done by callModule automatically.}

\item{data}{\code{reactivevalues} reactive data.table}

\item{optional_stats}{\code{character} optional statistics computed on numeric
data, default is "all". "pct_zero", "pct_NA", "mean", "median", "sd" are
always computed, possible values are "min", "max", "var", "ecart_interquartile",
"mode_max", "kurtosis", "skewness", "boxplot", "density".}

\item{nb_modal2show}{\code{integer} number of modalities to show 
for factor variables.}
}
\value{
shiny module
}
\description{
Shiny module to show descripive statistics on data
}
\examples{

\dontrun{
# In UI :
show_dataUI(id = "id")
# In Server, with data in a reactiveValues 
# for example
data <- reactiveValues(data = iris)
optional_stats = "all"
callModule(module = show_data, id = "id", data = reactive(data$data),
optional_stats = optional_stats)

## Examples apps
optional_stats <- "all"
run_example_app_show_data(optional_stats)
optional_stats <- "all"
run_example_app_filter_and_show_data(optional_stats)
}

}
\seealso{
\code{\link{filter_data}}
}
