% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_data.R
\name{show_dataUI}
\alias{show_dataUI}
\title{UI part of the module show_data}
\usage{
show_dataUI(id)
}
\arguments{
\item{id}{\code{character} An id that will be used to create a namespace}
}
\value{
UI page
}
\description{
This function has to be set in the UI part of a shiny application
}
\examples{
\dontrun{
# In UI :
show_dataUI(id = "id")
# In Server, with data in a reactiveValues 
# for example
data <- reactiveValues(data = iris)
optional_stats = "all"
callModule(module = show_data, id = "id", data = reactive(data$data),
optional_stats = optional_stats)

## For a complete example, you can run the function \\link{run_example_app_show_data}
optional_stats <- "all"
run_example_app_show_data(optional_stats)

## For another example which also uses the module \\link{filter_data}, you can
## run \\link{run_app_filter_and_show_data}
optional_stats <- "all"
run_app_filter_and_show_data(optional_stats)#' }

}
