% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_dt_table.R
\name{show_DT}
\alias{show_DT}
\title{Module d'affichage DT avec export. Partie SERVER}
\usage{
show_DT(input, output, session, data, dt, file_name, row.names = FALSE,
  server = TRUE)
}
\arguments{
\item{input, }{output, session : conventions shiny pour le modules}

\item{data}{: reactive. data.frame / data.table contenant les données}

\item{dt}{: reactive. dt (\code{datatable}) avec options à afficher}

\item{file_name}{: character. nom du fichier exporte (sans extension)}
}
\description{
Module d'affichage DT avec export. Partie SERVER
}
\examples{

# ui
ui = shiny::fluidPage(show_DT_UI("iris_module"))
server = function(input, output, session) {
  callModule(show_DT, "iris_module", reactive(iris), reactive(DT::datatable(iris)), 
   paste0("Iris_export", format(Sys.time(), format = "\%d\%m\%Y_\%H\%M\%S")))
}
shiny::shinyApp(ui = ui, server = server)
    
}
