% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_dt_table.R
\name{show_DT_UI}
\alias{show_DT_UI}
\alias{show_DT}
\title{Show & export table using DT}
\usage{
show_DT_UI(id, export = c("csv", "excel", "html"))

show_DT(
  input,
  output,
  session,
  data,
  dt,
  file_name,
  row.names = FALSE,
  server = TRUE,
  f_csv = NULL,
  f_xlsx = NULL,
  f_html = NULL
)
}
\arguments{
\item{id}{: character. id}

\item{export}{: character. Wanted export type ? Can be multiple}

\item{input}{Not a real parameter, should not be set manually. 
Done by callModule automatically.}

\item{output}{Not a real parameter, should not be set manually. 
Done by callModule automatically.}

\item{session}{Not a real parameter, should not be set manually. 
Done by callModule automatically.}

\item{data}{: reactive. \code{data.frame} / \code{data.table}}

\item{dt}{: reactive. dt (\code{datatable}) with options}

\item{file_name}{: character. Name of output file}

\item{row.names}{: \code{logical}}

\item{server}{: \code{logical}}

\item{f_csv}{: custom function(data, path_file) for .csv export}

\item{f_xlsx}{: custom function(data, path_file) for .xlsx export}

\item{f_html}{: custom function(data, path_file) for .html export}
}
\description{
Show & export table using DT
}
\examples{

\dontrun{
# ui
ui = shiny::fluidPage(
   show_DT_UI(
     id = "iris_module", 
     export = c("csv", "html")
   )
 )

server = function(input, output, session) {
  callModule(module = show_DT, id = "iris_module", 
   data = reactive(iris), 
   dt = reactive(DT::datatable(iris)), 
   file_name = paste0("Iris_export", format(Sys.time(), format = "\%d\%m\%Y_\%H\%M\%S")))
}
shiny::shinyApp(ui = ui, server = server)
  
# custom export function
f_csv = function(data, file){
  write.table(data, file, sep = ",", dec = ".")
}

server = function(input, output, session) {
  callModule(module = show_DT, id = "iris_module", 
   data = reactive(iris), 
   dt = reactive(DT::datatable(iris)), 
   file_name = paste0("Iris_export", format(Sys.time(), format = "\%d\%m\%Y_\%H\%M\%S")),
   f_csv = f_csv
   )
}
shiny::shinyApp(ui = ui, server = server)

}  

}
