% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_dt_table.R
\name{show_DT_UI}
\alias{show_DT_UI}
\title{Module d'affichage DT avec export. Partie UI}
\usage{
show_DT_UI(id)
}
\arguments{
\item{id}{: character. id du module}
}
\description{
Module d'affichage DT avec export. Partie UI
}
