% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showModal.R
\name{showModalServer}
\alias{showModalServer}
\title{Show Modal Server}
\usage{
showModalServer(
  id,
  ...,
  title = NULL,
  footer = modalButton("Dismiss"),
  size = c("m", "s", "l", "xl"),
  easyClose = FALSE,
  fade = TRUE
)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{...}{Named attributes to be applied to the button or link.}

\item{title}{An optional title for the dialog.}

\item{footer}{UI for footer. Use NULL for no footer.}

\item{size}{One of \code{"s"} for small, \code{"m"} (the default) for medium, or \code{"l"} for large.}

\item{easyClose}{If \code{TRUE}, the modal dialog can be dismissed by clicking outside
the dialog box, or be pressing the Escape key. If \code{FALSE} (the default),
the modal dialog can't be dismissed in those ways; instead it must be
dismissed by clicking on a \code{modalButton()}, or from a call to \code{\link[=removeModal]{removeModal()}}
on the server.}

\item{fade}{If \code{FALSE}, the modal dialog will have no fade-in animation
(it will simply appear rather than fade in to view).}
}
\value{
A show modal server
}
\description{
Create a show modal server.
}
\seealso{
\code{\link[=showModalUI]{showModalUI()}}
}
