% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showModal.R
\name{showModalUI}
\alias{showModalUI}
\title{Show Modal Action Button UI}
\usage{
showModalUI(id, label = "", icon = NULL, width = NULL, ...)
}
\arguments{
\item{id}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{icon}{An optional \code{\link[=icon]{icon()}} to appear on the button.}

\item{width}{The width of the input, e.g. \verb{400px}, or \verb{100\%}; see \code{\link[shiny:reexports]{shiny::validateCssUnit()}}.}

\item{...}{Named attributes to be applied to the button or link.}
}
\value{
A show modal action button UI
}
\description{
Create a show modal action button UI.
}
\seealso{
\code{\link[=showModalServer]{showModalServer()}}
}
