% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliderUI.R
\name{sliderInputUIs}
\alias{sliderInputUIs}
\title{Create tagList of sliderInputUI instances.}
\usage{
sliderInputUIs(ids, labels = NULL, values = NULL, minvals = NULL,
  maxvals = NULL, steps = NULL, widths = NULL, prefixes = NULL,
  postfixes = NULL, custom_css = NULL)
}
\description{
Create tagList of sliderInputUI instances.
}
