% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliderPeriod.R
\name{sliderPeriodServer}
\alias{sliderPeriodServer}
\title{Slider Period Input Control Server}
\usage{
sliderPeriodServer(
  id,
  choices,
  fun = function(x) format(x, "\%y.\%m"),
  invfun = function(x) as.Date(paste0(x, ".01"), "\%y.\%m.\%d"),
  selected = NULL,
  reverse = FALSE
)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{choices}{List of values to select from.}

\item{fun}{An expression function for \code{choices}}

\item{invfun}{An inverse function for \code{fun}}

\item{selected}{The initially selected value.}

\item{reverse}{A boolean value whether to reverse the choices or not}
}
\value{
The return value, if any, from executing the module server function
}
\description{
Create a slider period input control server.
}
\seealso{
\code{\link[=sliderPeriodUI]{sliderPeriodUI()}}
}
