% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slider.R
\name{sliderServer}
\alias{sliderServer}
\title{Slider Input Control Server}
\usage{
sliderServer(id, min = NULL, max = NULL, value = NULL, step = NULL)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{min, max}{The minimum and maximum values (inclusive) that can be
selected.}

\item{value}{The initial value of the slider, either a number, a date
(class Date), or a date-time (class POSIXt). A length one vector will
create a regular slider; a length two vector will create a double-ended
range slider. Must lie between \code{min} and \code{max}.}

\item{step}{Specifies the interval between each selectable value on the
slider. Either \code{NULL}, the default, which uses a heuristic to determine the
step size or a single number. If the values are dates, \code{step} is in days;
if the values are date-times, \code{step} is in seconds.}
}
\value{
A select input control server
}
\description{
Create a slider input control server.
}
\seealso{
\code{\link[=selectUI]{selectUI()}}
}
