% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliderText.R
\name{sliderTextServer}
\alias{sliderTextServer}
\title{Slider Text Input Control Server}
\usage{
sliderTextServer(id, choices, selected = NULL, reverse = FALSE)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{choices}{List of values to select from.}

\item{selected}{The initially selected value.}

\item{reverse}{A boolean value whether to reverse the choices or not}
}
\value{
The return value, if any, from executing the module server function
}
\description{
Create a slider text input control server.
}
\seealso{
\code{\link[=sliderTextUI]{sliderTextUI()}}
}
