% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliderText.R
\name{sliderTextUI}
\alias{sliderTextUI}
\title{Slider Text Input Control UI}
\usage{
sliderTextUI(
  id,
  label = "Period",
  choices = c(0, 1),
  selected = c(0, 1),
  hide_min_max = FALSE
)
}
\arguments{
\item{id}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}

\item{choices}{Default c(0, 1), list of values to select from.}

\item{selected}{Default C(0, 1), the initially selected value.}
}
\value{
A slider text input UI
}
\description{
Create a slider text input ui.
}
\seealso{
\code{\link[=sliderTextServer]{sliderTextServer()}}, \code{\link[=dynSliderTextServer]{dynSliderTextServer()}}
}
