% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_data.R
\name{summary_data_UI}
\alias{summary_data_UI}
\alias{summary_data}
\title{UI part of the module summary_data}
\usage{
summary_data_UI(id, titles = TRUE, subtitles = TRUE)

summary_data(
  input,
  output,
  session,
  data = NULL,
  optional_stats = "all",
  nb_modal2show = 3,
  columns_to_show = "all",
  show_warnings = FALSE,
  labels = list(title = "Descriptive statistics", num_var = "Numeric variables",
    date_var = "Date variables", factor_var = "Factor variables", message =
    "Calculation in progress...")
)
}
\arguments{
\item{id}{\code{character} An id that will be used to create a namespace
It returns three tables, one with statistics on numeric data, one with 
statistics on factor data and one with statistics on dates data}

\item{titles}{\code{logical} Add titles on UI ? Default to TRUE}

\item{subtitles}{\code{logical} Add subtitles on UI ? Default to TRUE}

\item{input}{Not a real parameter, should not be set manually. 
Done by callModule automatically.}

\item{output}{Not a real parameter, should not be set manually. 
Done by callModule automatically.}

\item{session}{Not a real parameter, should not be set manually. 
Done by callModule automatically.}

\item{data}{\code{data.frame/reactive} reactive data.table}

\item{optional_stats}{\code{character/reactive} optional statistics computed on numeric
data, default is "all". "pct_zero", "pct_NA", "mean", "median", "sd" are
always computed, possible values are "min", "max", "nb_valid", "var", "interquartile_range",
"mode_max", "kurtosis", "skewness", "boxplot", "density".}

\item{nb_modal2show}{\code{integer/reactive} number of modalities to show 
for factor variables.}

\item{columns_to_show}{\code{character/reactive} vector of column names you want to be
shown in the tables}

\item{show_warnings}{\code{logical/reactive} Show warnings ? (example compute Min. on all NAs)}

\item{labels}{\code{list/reactive} Title / subtitle / message
\itemize{
 \item{"title"}{ : Module title.}
 \item{"num_var "}{ : Subtitle for numeric variables.}
 \item{"date_var"}{ : Subtitle for date variables.}
 \item{"factor_var"}{ : Subtitle for factor variables.}
 \item{"message"}{ :  Message of the progress computing indicators}
}}
}
\value{
shiny module
}
\description{
Shiny module to show descripive statistics on data
}
\examples{

\dontrun{

ui = shiny::fluidPage(summary_data_UI(id = "id", titles = TRUE))

server = function(input, output, session) {
  data <- reactive({data})
  shiny::callModule(module = summary_data, id = "id", 
    data = data,
    optional_stats = "all"
   )
}

shiny::shinyApp(ui = ui, server = server)
     
## filter on stats 
ui = shiny::fluidPage(summary_data_UI(id = "id", titles = TRUE))
server = function(input, output, session) {
  shiny::callModule(module = summary_data, id = "id", data = iris,
    optional_stats = c("kurtosis", "density"),
    columns_to_show = c("Species", "Petal.Width", "Sepal.Width"))
}

shiny::shinyApp(ui = ui, server = server)   
          
## labels
ui = shiny::fluidPage(summary_data_UI(id = "id", titles = FALSE))
server = function(input, output, session) {
  shiny::callModule(module = summary_data, id = "id", data = iris,
    labels = list(title = "Statistiques descriptives", 
                  num_var = "Variables num\u00e9riques",
                  date_var = "Variables dates",
                  factor_var = "Variables Facteurs",
                  message = "Calcul en cours..."
    ))
}

shiny::shinyApp(ui = ui, server = server)   
                 
## Examples apps
run_example_app_summary_data()
run_example_app_filter_and_summary_data()
}

}
\seealso{
\code{\link{filter_data}}
}
