% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch.R
\name{switchServer}
\alias{switchServer}
\title{Slider Input Control Server}
\usage{
switchServer(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}
}
\value{
The return value, if any, from executing the module server function
}
\description{
Create a toggle switch control server.
}
\seealso{
\code{\link[=switchUI]{switchUI()}}
}
