% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch.R
\name{switchUI}
\alias{switchUI}
\title{Switch Input Control UI}
\usage{
switchUI(id, label = "Gender", onLabel = "Split", offLabel = "Merge")
}
\arguments{
\item{id}{The input slot that will be used to access the value.}

\item{label}{Display a text in the center of the switch.}

\item{onLabel}{Text on the left side of the switch (TRUE).}

\item{offLabel}{Text on the right side of the switch (FALSE).}
}
\value{
A switch Input UI
}
\description{
Create a toggle switch control ui.
}
\seealso{
\code{\link[=switchServer]{switchServer()}}
}
