% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{textServer}
\alias{textServer}
\title{Create a text input control Server}
\usage{
textServer(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}
}
\value{
The return value, if any, from executing the module server function
}
\description{
Create an input control server for entry of unstructured text values
}
\seealso{
\code{\link[=textUI]{textUI()}}
}
