% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{textUI}
\alias{textUI}
\title{Create a text input control UI}
\usage{
textUI(id, label = "", value = "")
}
\arguments{
\item{id}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}

\item{value}{Initial value.}
}
\value{
A text input UI
}
\description{
Create an input control ui for entry of unstructured text values
}
\seealso{
\code{\link[=textServer]{textServer()}}
}
