% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_counts}
\alias{transform_counts}
\title{Transform counts with the supplied function name}
\usage{
transform_counts(counts, transform_func)
}
\arguments{
\item{counts}{tibble with only numeric columns}

\item{transform_func}{character string to select the transform function}
}
\value{
a [tibble::tibble] conatining the transformed counts
}
\description{
Transform counts with the supplied function name
}
\examples{

transform_counts(rnaseqtools::counts, "raw")

transform_counts(rnaseqtools::counts, "zscore")

}
