% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadRNASeq.R
\name{uploadRNASeqApp}
\alias{uploadRNASeqApp}
\title{A test shiny app for the uploadRNASeq module}
\usage{
uploadRNASeqApp(testing = FALSE, debug = FALSE)
}
\value{
a [shiny::shinyApp()] object
}
\description{
`uploadRNASeqApp()` creates a small test app for testing the [uploadRNASeqInput()] and
[uploadRNASeqServer()] functions. A subset of the returned sample and count data.frames
are displayed in two [shiny::tableOutput()]s
}
\examples{
uploadRNASeqApp()
}
