% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadRNASeq.R
\name{uploadRNASeqOutput}
\alias{uploadRNASeqOutput}
\title{Create Output to display uploaded sample and count data}
\usage{
uploadRNASeqOutput(id)
}
\arguments{
\item{id}{namespace id for the UI components. Must match the id provided to the
[uploadRNASeqServer()] function.}
}
\value{
a [htmltools::tagList()] containing three [shiny::tableOutput()]s
and two [shiny::uiOutput()]s.
}
\description{
`uploadRNASeqOutput()` produces three table outputs for displaying
the data returned by the `uploadRNASeqServer()` function. The three
outputs are labelled samples, counts and metadata. There are also two
[shiny::uiOutput()]s to alert the user to missing data in
either the sample or count file
}
\examples{

uploadRNASeqOutput("rnaseqData")

}
