% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadRNASeq.R
\name{uploadRNASeqServer}
\alias{uploadRNASeqServer}
\title{Server function to upload sample and count data files}
\usage{
uploadRNASeqServer(
  id,
  test_data_paths = list(count_file = NULL, sample_file = NULL),
  testing = FALSE,
  debug = FALSE
)
}
\arguments{
\item{id}{namespace id for the UI components. Must match the id provided to the
[uploadRNASeqInput()] function.}

\item{test_data_paths}{list with 2 entries named count_file and sample_file
These should be paths to test data to use.}
}
\value{
a list containing two [shiny::reactive()] objects
* sampleInfo a data.frame of sample metadata
* counts a data.frame of RNAseq count data
}
\description{
`uploadRNASeqServer()` implements uploading a sample file and a count data
file. It also handles using the package test data.
}
\examples{

uploadRNASeqServer("rnaseqData")

}
