% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_data.R
\name{visualize_data}
\alias{visualize_data}
\title{server part of the module visualize_data}
\usage{
visualize_data(input, output, session, data = NULL)
}
\arguments{
\item{input}{Not a real parameter, should not be set manually. 
Done by callModule automatically.}

\item{output}{Not a real parameter, should not be set manually. 
Done by callModule automatically.}

\item{session}{Not a real parameter, should not be set manually. 
Done by callModule automatically.}

\item{data}{\code{reactivevalues} reactive data.table}
}
\value{
Server logic
}
\description{
This function has to be set in the Server part of a shiny application
It returns two table, one with statistics on numeric data and one with 
statistics on factor data
}
\examples{
\dontrun{
# In UI :
show_dataUI(id = "id")
# In Server, with data in a reactiveValues 
# for example
data <- reactiveValues(data = iris)
callModule(module = visualize, id = "id", data = reactive(data$data))

## For a complete example, you can run the function \\link{run_example_app_visualize_data}

run_example_app_visualize_data(run_example_app_visualize_data)

}

}
