% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_data.R
\name{visualize_dataUI}
\alias{visualize_dataUI}
\alias{visualize_data}
\title{UI part of the module visualize_data}
\usage{
visualize_dataUI(id, titles = TRUE)

visualize_data(input, output, session, data = NULL)
}
\arguments{
\item{id}{\code{character} An id that will be used to create a namespace}

\item{titles}{\code{logical} Add titles on UI ? Default to TRUE}

\item{input}{Not a real parameter, should not be set manually. 
Done by callModule automatically.}

\item{output}{Not a real parameter, should not be set manually. 
Done by callModule automatically.}

\item{session}{Not a real parameter, should not be set manually. 
Done by callModule automatically.}

\item{data}{\code{reactivevalues} reactive data.table}
}
\value{
UI page
}
\description{
This function has to be set in the UI part of a shiny application
}
\examples{
\dontrun{

ui = shiny::fluidPage(visualize_dataUI(id = "id", titles = TRUE))
server = function(input, output, session) {
  data <- reactiveValues(data = iris)
  shiny::callModule(module = visualize_data, id = "id", data = reactive(data$data))
}

shiny::shinyApp(ui = ui, server = server)

## Example apps
run_example_app_visualize_data()
}


}
