% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_data.R
\name{visualize_dataUI}
\alias{visualize_dataUI}
\title{UI part of the module visualize_data}
\usage{
visualize_dataUI(id)
}
\arguments{
\item{id}{\code{character} An id that will be used to create a namespace}
}
\value{
UI page
}
\description{
This function has to be set in the UI part of a shiny application
}
