% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xVyScatterPlot.R
\name{xVyScatterplotApp}
\alias{xVyScatterplotApp}
\title{A test shiny app for the xVyScatterplot module}
\usage{
xVyScatterplotApp(debug = FALSE)
}
\value{
a [shiny::shinyApp()] object
}
\description{
`xVyScatterplotApp()` creates a small test app for testing the [xVyScatterplotInput()] and
[xVyScatterplotServer()] functions. A subset of the returned sample and count data.frames
are displayed in two [shiny::tableOutput()]s
}
\examples{
xVyScatterplotApp()
}
