% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xVyScatterPlot.R
\name{xVyScatterplotInput}
\alias{xVyScatterplotInput}
\title{Create UI components to select the variables to plot against each other}
\usage{
xVyScatterplotInput(id, x_label = "X variable", y_label = "Y variable")
}
\arguments{
\item{id}{namespace id for the UI components. Must match the id provided to the
[xVyScatterplotServer()] function.}

\item{x_label}{character Label for X variable select input}

\item{y_label}{character Label for Y variable select input}
}
\value{
a [htmltools::tagList()] containing two [shiny::selectizeInput()]
controls and one [shiny::checkboxInput()]
}
\description{
`xVyScatterplotInput()` produces the select inputs for selecting the
variables to plot against each other
}
\examples{

xVyScatterplotInput("countData")

}
