% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xVyScatterPlot.R
\name{xVyScatterplotOutput}
\alias{xVyScatterplotOutput}
\title{Create Output to display the Scatterplot}
\usage{
xVyScatterplotOutput(id)
}
\arguments{
\item{id}{namespace id for the UI components. Must match the id provided to the
[xVyScatterplotServer()] function.}
}
\value{
a [htmltools::tagList()] containing three [shiny::plotOutput()]s
}
\description{
`xVyScatterplotOutput()` produces a PlotOutput element for the scatterplot
}
\examples{

xVyScatterplotOutput("countData")

}
