% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xVyScatterPlot.R
\name{xVyScatterplotServer}
\alias{xVyScatterplotServer}
\title{Server function to produce an xy scatter plot}
\usage{
xVyScatterplotServer(id, data = NULL, debug = FALSE)
}
\arguments{
\item{id}{namespace id for the UI components. Must match the id provided to the
[xVyScatterplotInput()] function.}

\item{data}{data.frame for plotting. Expected in long format with a column
named `variable` and one named `value`}
}
\value{
a list invisibly containing one [shiny::reactive()] object
* plot_data a data.frame of plotted data
}
\description{
`xVyScatterplotServer()` takes a data frame and produces a scatterplot
based on the variables selected
}
\examples{

xVyScatterplotServer("rnaseqData")

}
