% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countPlot.R
\name{countPlotInput}
\alias{countPlotInput}
\title{Create UI components to display a count plot for a selected gene}
\usage{
countPlotInput(id)
}
\arguments{
\item{id}{namespace id for the UI components. Must match the id provided to the
[countPlotInput()] and [countPlotOutput()] functions.}
}
\value{
a [htmltools::tagList()] containing a [shiny::selectInput()] object
}
\description{
`countPlotInput()` produces a [shiny::selectInput()] control for choosing a gene
to display as a count plot.
}
\examples{

countPlotOutput("rnaseqData")

}
