% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countPlot.R
\name{countPlotServer}
\alias{countPlotServer}
\title{Server function to create a count plot for a given gene}
\usage{
countPlotServer(
  id,
  counts = NULL,
  sample_info = NULL,
  gene_metadata = NULL,
  debug = FALSE
)
}
\arguments{
\item{id}{namespace id for the UI components. Must match the id provided to the
[countPlotInput()] function.}

\item{counts}{a reactive counts object. Should contain only numeric columns}

\item{sample_info}{a reactive object. Represents the samples and associated
metadata}

\item{gene_metadata}{a reactive object. Contains the metadata for the genes
present in the counts object.}

\item{debug}{Turn on debugging message statements}
}
\value{
a [shiny::reactive()] object which is the count plot
}
\description{
`countPlotServer()` creates a count plot for a given gene from the counts and
sample information supplied. The plot is a [ggplot2] plot object made using
[biovisr::scatterplot_with_fill_and_shape()]. The plot is rendered to the
count_plot plotOutput from the `countPlotOutput()` function.
}
\examples{

countPlotServer("rnaseq", counts = reactive(rnaseqtools::counts[1:10, 1:20]),
sample_info = reactive(rnaseqtools::sampleInfo[1:20,]),
gene_metadata = reactive(rnaseqtools::gene_metadata[1:10,]))

}
