% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countPlot.R
\name{create_count_plot}
\alias{create_count_plot}
\title{Make scatterplot of normalised counts}
\usage{
create_count_plot(
  count_data,
  x_var,
  colour_var,
  colour_palette,
  shape_var = NULL,
  shape_palette = c(21:24),
  debug = FALSE
)
}
\arguments{
\item{count_data}{[tibble::tibble()] of count data joined to sample info}

\item{x_var}{character column name to use as the X variable}

\item{colour_var}{character column name to use as the colour variable}

\item{colour_palette}{character named vector of colours}

\item{shape_var}{character column name to use as the shape variable.
It can be NULL in which case the points are all circles}

\item{shape_palette}{numeric vector of numbers for the shapes}

\item{debug}{logical print debugging info}
}
\value{
a [ggplot2::ggplot()] object
}
\description{
Make scatterplot of normalised counts
}
\examples{

create_count_plot(counts_for_gene, x_var = 'condition', colour_var = 'treatment',
colour_palette = c('yellow', 'blue', 'red'), shape_var = 'timepoint')

}
