% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countPlot.R
\name{create_fill_palette}
\alias{create_fill_palette}
\title{Make a vector of colours based on a character vector}
\usage{
create_fill_palette(fill_vec)
}
\arguments{
\item{fill_vec}{factor the column to based the colour palette on}
}
\value{
character named vector of colours for each level of the input vector
}
\description{
Make a vector of colours based on a character vector
}
\examples{

create_fill_palette(counts_for_gene$treatment)

}
