% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countPlot.R
\name{create_shape_palette}
\alias{create_shape_palette}
\title{Make a vector of numbers representing shapes based on a character vector}
\usage{
create_shape_palette(shape_vec)
}
\arguments{
\item{shape_vec}{factor the column to based the shape palette on}
}
\value{
integer vector of numbers that represent plotting shapes
see \link{https://ggplot2.tidyverse.org/reference/scale_shape.html}
}
\description{
Make a vector of numbers representing shapes based on a character vector
}
\examples{

create_shape_palette(counts_for_gene$timepoint)

create_shape_palette(factor(letters[1:6]))

}
