% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTableModule.R
\name{dataTableServer}
\alias{dataTableServer}
\title{the server for a DataTable shiny module}
\usage{
dataTableServer(
  id,
  input,
  output,
  session,
  tbl,
  selectionPolicy = reactive("multiple"),
  wrapLongTextInCells = reactive(TRUE),
  searchString = reactive(""),
  rownames.to.display = reactive("all")
)
}
\arguments{
\item{input}{enviroment provided by shiny}

\item{output}{enviroment provided by shiny}

\item{session}{enviroment provided by shiny}

\item{tbl}{data.frame}

\item{selectionPolicy}{character string, "none", "single", or "multiple"}

\item{wrapLongTextInCells}{logical, TRUE or FALSE}

\item{searchString}{character string, selects all rows with this, default "" (no search)}

\item{rownames.to.display}{character vector, default "all",}
}
\description{
the server for a DataTable shiny module
}
