% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTableModule.R
\name{dataTableServer}
\alias{dataTableServer}
\title{the server for a DataTable shiny module}
\usage{
dataTableServer(
  input,
  output,
  session,
  tbl,
  selectionPolicy = reactive("single"),
  pageLength = reactive(5),
  visibleRows = reactive(head(rownames(tbl))),
  selectedRows = reactive(NULL),
  searchTerm = reactive(NULL),
  wrapLongTextInCells = reactive(TRUE)
)
}
\arguments{
\item{input}{enviroment provide by shiny}

\item{output}{enviroment provide by shiny}

\item{session}{enviroment provide by shiny}

\item{tbl}{data.frame}

\item{selectionPolicy}{character string, single, multiple or none}

\item{pageLength}{integer typically 5, 10, 25 or 50}

\item{visibleRows}{"all", "none", or a list of rownames}

\item{selectedRows}{character string, reactive(NULL)  by default}

\item{searchTerm}{character string, "" by default}

\item{wrapLongTextInCells}{logcial default TRUE}
}
\description{
the server for a DataTable shiny module
}
