% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicators.R
\name{.get_dates_indicators}
\alias{.get_dates_indicators}
\title{get stats indicator on dates data}
\usage{
.get_dates_indicators(data, dates_vars, optional_stats)
}
\arguments{
\item{data}{\code{data.table}}

\item{dates_vars}{\code{character}}
}
\value{
DT with statistics on dates data
}
\description{
return statistics from dates data, is called by 
\link{get_dt_num_dt_fac}
}
