% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicators.R
\name{.get_factor_indicators}
\alias{.get_factor_indicators}
\title{get stats indicator on factor data}
\usage{
.get_factor_indicators(
  data,
  fact_vars,
  nb_modal2show,
  optional_stats,
  keep_dataframe = TRUE,
  keep_datatable = TRUE
)
}
\arguments{
\item{data}{\code{data.table}}

\item{fact_vars}{\code{character}}

\item{nb_modal2show}{\code{integer}}

\item{keep_dataframe}{\code{logical} data.frame in output ?}

\item{keep_datatable}{\code{logical} datatable (DT) in output ?}
}
\value{
DT with statistics on factor data
}
\description{
return statistics from factor data, is called by 

\link{get_dt_num_dt_fac}
}
