% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicators.R
\name{.get_indicators}
\alias{.get_indicators}
\title{get stats indicator on numeric column}
\usage{
.get_indicators(data, var, absolute = FALSE, optional_stats)
}
\arguments{
\item{data}{\code{data.table}}

\item{var}{\code{character}}

\item{absolute}{\code{logical}}

\item{optional_stats}{\code{character}}
}
\value{
data.table with statistics on numeric data
}
\description{
return statistics from numeric data, is called by 
\link{.get_stat_indicators}
}
