% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicators.R
\name{.get_stat_indicators}
\alias{.get_stat_indicators}
\title{get stats indicator on numeric data}
\usage{
.get_stat_indicators(
  data,
  vars,
  optional_stats,
  keep_dataframe = TRUE,
  keep_datatable = TRUE,
  progress = NULL
)
}
\arguments{
\item{data}{\code{data.frame}, \code{data.table}}

\item{vars}{\code{character}}

\item{optional_stats}{\code{character} optional statistics computed on data,
you can look at \link{show_data} for more information.}

\item{keep_dataframe}{\code{logical} data.frame in output ?}

\item{keep_datatable}{\code{logical} datatable (DT) in output ?}

\item{progress}{Optionnal. Shiny progress bar.}
}
\value{
DT with statistics on numeric data
}
\description{
return statistics from numeric data, is called by 
\link{get_dt_num_dt_fac}
}
