% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentalMeasuresModule.R
\name{ExperimentalMeasuresServer}
\alias{ExperimentalMeasuresServer}
\title{the server for a ExperimentalMesures shiny module}
\usage{
ExperimentalMeasuresServer(id, tbl)
}
\arguments{
\item{id}{html element identifier}

\item{tbl, }{a data.frame}
}
\description{
the server for a ExperimentalMesures shiny module
}
