% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentalMeasuresModule.R
\name{ExperimentalMeasuresServer}
\alias{ExperimentalMeasuresServer}
\title{the server for a ExperimentalMesures shiny module}
\usage{
ExperimentalMeasuresServer(id, tbl)
}
\arguments{
\item{tbl, }{a data.frame}

\item{input}{enviroment provide by shiny}

\item{output}{enviroment provide by shiny}

\item{session}{enviroment provide by shiny}
}
\description{
the server for a ExperimentalMesures shiny module
}
