% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messageBoxModule.R
\name{messageBoxUI}
\alias{messageBoxUI}
\title{the UI for a very simple message box module}
\usage{
messageBoxUI(
  id,
  title,
  boxWidth = 200,
  boxHeight = 30,
  fontSize = 20,
  fontColor = "black",
  backgroundColor = "beige"
)
}
\arguments{
\item{id}{character string, the html document's widget id}

\item{title}{character string, title for the box}

\item{boxWidth}{integer pixels, 200 by default}

\item{boxHeight}{integer pixels, 30 by default}

\item{fontSize}{integer pixels, 20 by default}

\item{fontColor}{character string, uses standard CSS naming conventions, "black" by default}

\item{backgroundColor}{character string, uses standard CSS naming conventions, "beige" by default}
}
\description{
the UI for a very simple message box module
}
