% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliderPeriod.R
\name{sliderPeriodUI}
\alias{sliderPeriodUI}
\title{Slider Period Input Control UI}
\usage{
sliderPeriodUI(id, label = "Period")
}
\arguments{
\item{id}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}
}
\value{
A slider period input UI
}
\description{
Create a slider period input ui.
}
\seealso{
\code{\link[=sliderPeriodServer]{sliderPeriodServer()}}, \code{\link[=dynSliderPeriodServer]{dynSliderPeriodServer()}}
}
