% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadRNASeq.R
\name{uploadRNASeqInput}
\alias{uploadRNASeqInput}
\title{Create UI components to upload sample and count files}
\usage{
uploadRNASeqInput(id)
}
\arguments{
\item{id}{namespace id for the UI components. Must match the id provided to the
[uploadRNASeqServer()] function.}
}
\value{
a [htmltools::tagList()] containing two [shiny::fileInput()] controls and
a [shiny::checkboxInput()].
}
\description{
`uploadRNASeqInput()` produces the buttons for uploading sample and count data
files. It also provides a checkbox to use the package test data. If sample and
count files are subsequently uploaded, this is used instead of the test data.
}
\examples{

uploadRNASeqInput("rnaseqData")

}
