% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_indicators.R
\name{vis_indicators}
\alias{vis_indicators}
\title{Shiny module server-like fun to display indicators on a data.table.}
\usage{
vis_indicators(
  input,
  output,
  session,
  data,
  col_obs,
  col_fit,
  col_date = NULL,
  indicators
)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny input}

\item{session}{shiny input}

\item{data}{\code{data.frame} / \code{data.table}. Table on which indicators will be computed.}

\item{col_obs}{\code{character}. Column name for observed values.}

\item{col_fit}{\code{character}. Column name for fitted values.}

\item{col_date}{\code{character} (NULL). Column name for date values.}

\item{indicators}{\code{characters}. Indicators to be computed, amongst: ("rmse", "mae", "mape", "mape_star").}
}
\value{
shiny module.
}
\description{
Shiny module server-like fun to display indicators on a data.table.
}
\examples{
\dontrun{\donttest{

library(data.table)

data <- data.table(obs = runif(100, 1, 10))
data[, fit := obs + rnorm(100, 0, 10)]
data[, date := seq(as.POSIXct("2019-10-07 00:00:00 UTC", tz = "UTC"), 
    as.POSIXct("2019-10-11 03:00:00 UTC", tz = "UTC"), by = 60*60)]
data[, by_quali := factor(sample(rep(1:10, 10)))]
data[, by_quanti := runif(100, 1, 20)]

col_obs <- "obs"
col_fit <- "fit"
indicators <- c("rmse", "mae", "mape", "mape_e")

ui <- shiny::fluidPage(vis_indicators_UI("my_id", data, col_obs, col_fit))
server <- function(input, output, session) {
  callModule(vis_indicators, "my_id", data, col_obs, col_fit, indicators)
}
shiny::shinyApp(ui = ui, server = server)

}}
}
